extendr_api

Attribute Macro extendr

#[extendr]

The #[extendr]-macro may be placed on three items

fn for wrapped rust-functions, see extendr-fn

impl-blocks, see extendr-impl

There is also extendr_module, which is used for defining what rust wrapped items should be visible to the surrounding R-package.