Enum extendr_api::Rany

pub enum Rany<'a> {

    Null(&'a Robj),
    Symbol(&'a Symbol),
    Pairlist(&'a Pairlist),
    Function(&'a Function),
    Environment(&'a Environment),
    Promise(&'a Promise),
    Language(&'a Language),
    Special(&'a Primitive),
    Builtin(&'a Primitive),
    Rstr(&'a Rstr),
    Logical(&'a Robj),
    Integer(&'a Integers),
    Real(&'a Doubles),
    Complex(&'a Robj),
    String(&'a Robj),
    Dot(&'a Robj),
    Any(&'a Robj),
    List(&'a List),
    Expression(&'a Expression),
    Bytecode(&'a Robj),
    ExternalPtr(&'a Robj),
    WeakRef(&'a Robj),
    Raw(&'a Raw),
    S4(&'a S4),
    Unknown(&'a Robj),

}

Enum use to unpack R objects into their specialist wrappers. Todo: convert all Robj types to wrappers. Note: this only works if the wrappers are all just SEXPs.

Variants

Null(&'a Robj)

Tuple Fields of Null

0: &'a Robj

Symbol(&'a Symbol)

Tuple Fields of Symbol

0: &'a Symbol

Pairlist(&'a Pairlist)

Tuple Fields of Pairlist

0: &'a Pairlist

Function(&'a Function)

Tuple Fields of Function

0: &'a Function

Environment(&'a Environment)

Tuple Fields of Environment

0: &'a Environment

Promise(&'a Promise)

Tuple Fields of Promise

0: &'a Promise

Language(&'a Language)

Tuple Fields of Language

0: &'a Language

Special(&'a Primitive)

Tuple Fields of Special

0: &'a Primitive

Builtin(&'a Primitive)

Tuple Fields of Builtin

0: &'a Primitive

Rstr(&'a Rstr)

Tuple Fields of Rstr

0: &'a Rstr

Logical(&'a Robj)

Tuple Fields of Logical

0: &'a Robj

Integer(&'a Integers)

Tuple Fields of Integer

0: &'a Integers

Real(&'a Doubles)

Tuple Fields of Real

0: &'a Doubles

Complex(&'a Robj)

Tuple Fields of Complex

0: &'a Robj

String(&'a Robj)

Tuple Fields of String

0: &'a Robj

Dot(&'a Robj)

Tuple Fields of Dot

0: &'a Robj

Any(&'a Robj)

Tuple Fields of Any

0: &'a Robj

List(&'a List)

Tuple Fields of List

0: &'a List

Expression(&'a Expression)

Tuple Fields of Expression

0: &'a Expression

Bytecode(&'a Robj)

Tuple Fields of Bytecode

0: &'a Robj

ExternalPtr(&'a Robj)

Tuple Fields of ExternalPtr

0: &'a Robj

WeakRef(&'a Robj)

Tuple Fields of WeakRef

0: &'a Robj

Raw(&'a Raw)

Tuple Fields of Raw

0: &'a Raw

S4(&'a S4)

Tuple Fields of S4

0: &'a S4

Unknown(&'a Robj)

Tuple Fields of Unknown

0: &'a Robj