Enum extendr_api::Robj

pub enum Robj {
    // some variants omitted
}

Wrapper for an R S-expression pointer (SEXP).

Create R objects from rust types and iterators:

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

// Different ways of making integer scalar 1.
let non_na : Option<i32> = Some(1);
let a : Robj = vec![1].into();
let b = r!(1);
let c = r!(vec![1]);
let d = r!(non_na);
let e = r!([1]);
assert_eq!(a, b);
assert_eq!(a, c);
assert_eq!(a, d);
assert_eq!(a, e);

// Different ways of making boolean scalar TRUE.
let a : Robj = true.into();
let b = r!(TRUE);
assert_eq!(a, b);

// Create a named list
let a = list!(a = 1, b = "x");
assert_eq!(a.len(), 2);

// Use an iterator (like 1:10)
let a = r!(1 ..= 10);
assert_eq!(a, r!([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]));

// Use an iterator (like (1:10)[(1:10) %% 3 == 0])
let a = (1 ..= 10).filter(|v| v % 3 == 0).collect_robj();
assert_eq!(a, c!(3, 6, 9));

Use iterators to get the contents of R objects.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let a : Robj = c!(1, 2, 3, 4, 5);
let iter = a.integer_iter().unwrap();
let robj = iter.filter(|&&x| x < 3).collect_robj();
assert_eq!(robj, c!(1, 2));

Convert to/from Rust vectors.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let a : Robj = r!(vec![1., 2., 3., 4.]);
let b : Vec<f64> = a.as_real_vector().unwrap();
assert_eq!(a.len(), 4);
assert_eq!(b, vec![1., 2., 3., 4.]);

Iterate over names and values.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let abc = list!(a = 1, b = "x", c = vec![1, 2]);
let names : Vec<_> = abc.names().unwrap().collect();
let names_and_values : Vec<_> = abc.named_list_iter().unwrap().collect();
assert_eq!(names, vec!["a", "b", "c"]);
assert_eq!(names_and_values, vec![("a", r!(1)), ("b", r!("x")), ("c", r!(vec![1, 2]))]);

NOTE: as much as possible we wish to make this object safe (ie. no segfaults).

If you avoid using unsafe functions it is more likely that you will avoid panics and segfaults. We will take great trouble to ensure that this is true.

Implementations

impl Robj

pub fn pairlist_iter(&self) -> Option<ListIter>

Get an iterator over a pairlist, lang or ... object. Note: we plan to combine list iterators into one.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let mut robj = lang!("+", 1, 2);
let objects : Vec<_> = robj.pairlist_iter().unwrap().collect();
assert_eq!(objects, vec![r!(Symbol("+")), r!(1), r!(2)])

pub fn list_iter(&self) -> Option<VecIter>

Get an iterator over an unnamed list. Note: we plan to combine list iterators into one.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let mut robj = list!(1, 2, 3);
let objects : Vec<_> = robj.list_iter().unwrap().collect();
assert_eq!(objects, vec![r!(1), r!(2), r!(3)])

pub fn str_iter(&self) -> Option<StrIter>

Get an iterator over a string vector. Returns None if the object is not a string vector but works for factors.

use extendr_api::*;

extendr_engine::start_r();

let obj = Robj::from(vec!["a", "b", "c"]);
assert_eq!(obj.str_iter().unwrap().collect::<Vec<_>>(), vec!["a", "b", "c"]);

let factor = factor!(vec!["abcd", "def", "fg", "fg"]);
assert_eq!(factor.levels().unwrap().collect::<Vec<_>>(), vec!["abcd", "def", "fg"]);
assert_eq!(factor.as_integer_vector().unwrap(), vec![1, 2, 3, 3]);
assert_eq!(factor.str_iter().unwrap().collect::<Vec<_>>(), vec!["abcd", "def", "fg", "fg"]);
assert_eq!(factor.str_iter().unwrap().collect::<Vec<_>>(), vec!["abcd", "def", "fg", "fg"]);

impl Robj

The following impls wrap specific Rinternals.h functions.

pub fn is_null(&self) -> bool

Return true if this is the null object.

pub fn is_symbol(&self) -> bool

Return true if this is a symbol.

pub fn is_logical(&self) -> bool

Return true if this is a boolean (logical) vector

pub fn is_real(&self) -> bool

Return true if this is a real (f64) vector.

pub fn is_complex(&self) -> bool

Return true if this is a complex vector.

pub fn is_expression(&self) -> bool

Return true if this is an expression.

pub fn is_environment(&self) -> bool

Return true if this is an environment.

pub fn is_string(&self) -> bool

Return true if this is a string.

pub fn is_object(&self) -> bool

Return true if this is an object.

pub fn get_current_srcref(val: i32) -> Robj

Get the source ref.

pub fn get_src_filename(&self) -> Robj

Get the source filename.

pub fn as_char(&self) -> Robj

Convert to a string vector.

pub fn coerce_vector(&self, sexptype: u32) -> Robj

Convert to vectors of many kinds.

pub fn pair_to_vector_list(&self) -> Robj

Convert a pairlist (LISTSXP) to a vector list (VECSXP).

pub fn vector_to_pair_list(&self) -> Robj

Convert a vector list (VECSXP) to a pair list (LISTSXP)

pub fn as_character_factor(&self) -> Robj

Convert a factor to a string vector.

pub fn alloc_matrix(sexptype: SEXPTYPE, rows: i32, cols: i32) -> Robj

Allocate a matrix object.

pub fn duplicate(&self) -> Self

Compatible way to duplicate an object. Use obj.clone() instead for Rust compatibility.

pub fn find_function<S>(&self, symbol: S) -> Result<Robj, AnyError> where
Robj: From<S>,

Find a function in an environment ignoring other variables.

   use extendr_api::*;
   extendr_engine::start_r();

   R!(my_fun <- function() {});
   let my_fun = global_env().find_function(Symbol("my_fun")).unwrap();
   assert_eq!(my_fun.is_function(), true);

pub fn ncols(&self) -> usize

Number of columns of a matrix

pub fn nrows(&self) -> usize

Number of rows of a matrix

pub fn xlengthgets(&self, new_len: usize) -> Result<Robj, AnyError>

Read-only access to attribute list. Copy a vector and resize it. See. https://github.com/hadley/r-internals/blob/master/vectors.md

pub fn alloc_vector(sexptype: u32, len: usize) -> Robj

Allocated an owned object of a certain type.

pub fn conformable(a: &Robj, b: &Robj) -> bool

Return true if two arrays have identical dims.

pub fn elt(&self, index: usize) -> Robj

Borrow an element from a list.

pub fn is_array(&self) -> bool

Return true if this is an array.

pub fn is_factor(&self) -> bool

Return true if this is factor.

pub fn is_frame(&self) -> bool

Return true if this is a data frame.

pub fn is_function(&self) -> bool

Return true if this is a function.

pub fn is_integer(&self) -> bool

Return true if this is an integer vector.

pub fn is_language(&self) -> bool

Return true if this is a language object.

pub fn is_list(&self) -> bool

Return true if this is a vector list.

pub fn is_matrix(&self) -> bool

Return true if this is a matrix.

pub fn is_new_list(&self) -> bool

Return true if this is a vector list or null.

pub fn is_number(&self) -> bool

Return true if this is a numeric vector but not a factor.

pub fn is_pair_list(&self) -> bool

Return true if this is a pairlist.

pub fn is_primitive(&self) -> bool

Return true if this is a primitive function.

pub fn is_ts(&self) -> bool

Return true if this is a time series vector (see tsp).

pub fn is_user_binop(&self) -> bool

Return true if this is a user defined binop.

pub fn is_valid_string(&self) -> bool

Return true if this is a valid string.

pub fn is_valid_string_f(&self) -> bool

Return true if this is a valid string.

pub fn is_vector(&self) -> bool

Return true if this is a vector.

pub fn is_vector_atomic(&self) -> bool

Return true if this is an atomic vector.

pub fn is_vector_list(&self) -> bool

Return true if this is a vector list.

pub fn is_vectorizable(&self) -> bool

Return true if this is can be made into a vector.

impl Robj

pub fn len(&self) -> usize

Get the extended length of the object.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let a : Robj = r!(vec![1., 2., 3., 4.]);
assert_eq!(a.len(), 4);

pub fn is_na(&self) -> bool

Is this object is an NA scalar? Works for character, integer and numeric types.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

assert_eq!(r!(NA_INTEGER).is_na(), true);
assert_eq!(r!(NA_REAL).is_na(), true);
assert_eq!(r!(NA_STRING).is_na(), true);

pub fn as_integer_slice<'a>(&self) -> Option<&'a [i32]>

Get a read-only reference to the content of an integer vector.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([1, 2, 3]);
assert_eq!(robj.as_integer_slice().unwrap(), [1, 2, 3]);

pub fn integer_iter<'a>(&self) -> Option<IntegerIter<'a>> where
Self: 'a,

Get an iterator over integer elements of this slice.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([1, 2, 3]);
let mut tot = 0;
for val in robj.integer_iter().unwrap() {
  tot += val;
}
assert_eq!(tot, 6);

pub fn as_integer_vector(&self) -> Option<Vec<i32>>

Get a Vec copied from the object.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([1, 2, 3]);
assert_eq!(robj.as_integer_slice().unwrap(), vec![1, 2, 3]);

pub fn as_logical_slice(&self) -> Option<&[Bool]>

Get a read-only reference to the content of a logical vector using the tri-state Bool. Returns None if not a logical vector.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([TRUE, FALSE, NA_LOGICAL]);
assert_eq!(robj.as_logical_slice().unwrap(), [TRUE, FALSE, NA_LOGICAL]);

pub fn as_logical_vector(&self) -> Option<Vec<Bool>>

Get a Vec copied from the object using the tri-state Bool. Returns None if not a logical vector.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([TRUE, FALSE, NA_LOGICAL]);
assert_eq!(robj.as_logical_vector().unwrap(), vec![TRUE, FALSE, NA_LOGICAL]);

pub fn logical_iter(&self) -> Option<LogicalIter<'_>>

Get an iterator over logical elements of this slice.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([TRUE, FALSE, NA_LOGICAL]);
let (mut nt, mut nf, mut nna) = (0, 0, 0);
for val in robj.logical_iter().unwrap() {
  match *val {
    TRUE => nt += 1,
    FALSE => nf += 1,
    NA_LOGICAL => nna += 1,
    _ => ()
  }
}
assert_eq!((nt, nf, nna), (1, 1, 1));

pub fn as_real_slice(&self) -> Option<&[f64]>

Get a read-only reference to the content of a double vector. Note: the slice may contain NaN or NA values. We may introduce a "Real" type to handle this like the Bool type.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([Some(1.), None, Some(3.)]);
let mut tot = 0.;
for val in robj.as_real_slice().unwrap() {
  if !val.is_na() {
    tot += val;
  }
}
assert_eq!(tot, 4.);

pub fn real_iter(&self) -> Option<RealIter<'_>>

Get an iterator over real elements of this slice.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([1., 2., 3.]);
let mut tot = 0.;
for val in robj.real_iter().unwrap() {
  if !val.is_na() {
    tot += val;
  }
}
assert_eq!(tot, 6.);

pub fn as_real_vector(&self) -> Option<Vec<f64>>

Get a Vec copied from the object.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([1., 2., 3.]);
assert_eq!(robj.as_real_vector().unwrap(), vec![1., 2., 3.])

pub fn as_raw_slice(&self) -> Option<&[u8]>

Get a read-only reference to the content of an integer or logical vector.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let robj = r!([1_u8, 2, 3]);
assert_eq!(robj.as_raw_slice().unwrap(), [1_u8, 2, 3]);

pub fn as_integer_slice_mut(&mut self) -> Option<&mut [i32]>

Get a read-write reference to the content of an integer or logical vector. Note that rust slices are 0-based so slice[1] is the middle value.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let mut robj = r!([1, 2, 3]);
let slice : & mut [i32] = robj.as_integer_slice_mut().unwrap();
slice[1] = 100;
assert_eq!(robj, r!([1, 100, 3]));

pub fn as_real_slice_mut(&mut self) -> Option<&mut [f64]>

Get a read-write reference to the content of a double vector. Note that rust slices are 0-based so slice[1] is the middle value.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let mut robj = r!([1.0, 2.0, 3.0]);
let slice = robj.as_real_slice_mut().unwrap();
slice[1] = 100.0;
assert_eq!(robj, r!([1.0, 100.0, 3.0]));

pub fn as_raw_slice_mut(&mut self) -> Option<&mut [u8]>

Get a read-write reference to the content of a raw vector.

use extendr_api::*;        // Put API in scope.
extendr_engine::start_r(); // Start test environment.

let mut robj = r!([1_u8, 2, 3]);
let slice = robj.as_raw_slice_mut().unwrap();
slice[1] = 100;
assert_eq!(robj, r!([1_u8, 100, 3]));

pub fn as_string_vector(&self) -> Option<Vec<String>>

Get a vector of owned strings. Owned strings have long lifetimes, but are much slower than references.

   use extendr_api::*;
   extendr_engine::start_r();
   let robj1 = Robj::from("xyz");
   assert_eq!(robj1.as_string_vector(), Some(vec!["xyz".to_string()]));
   let robj2 = Robj::from(1);
   assert_eq!(robj2.as_string_vector(), None);

pub fn as_str_vector(&self) -> Option<Vec<&str>>

Get a vector of string references. String references (&str) are faster, but have short lifetimes.

   use extendr_api::*;
   extendr_engine::start_r();
   let robj1 = Robj::from("xyz");
   assert_eq!(robj1.as_str_vector(), Some(vec!["xyz"]));
   let robj2 = Robj::from(1);
   assert_eq!(robj2.as_str_vector(), None);

pub fn as_str(&self) -> Option<&str>

Get a read-only reference to a char, symbol or string type.

   use extendr_api::*;
   extendr_engine::start_r();
   let robj1 = Robj::from("xyz");
   let robj2 = Robj::from(1);
   assert_eq!(robj1.as_str(), Some("xyz"));
   assert_eq!(robj2.as_str(), None);

pub fn as_integer(&self) -> Option<i32>

Get a scalar integer.

   use extendr_api::*;
   extendr_engine::start_r();
   let robj1 = Robj::from("xyz");
   let robj2 = Robj::from(1);
   let robj3 = Robj::from(NA_INTEGER);
   assert_eq!(robj1.as_integer(), None);
   assert_eq!(robj2.as_integer(), Some(1));
   assert_eq!(robj3.as_integer(), None);

pub fn as_real(&self) -> Option<f64>

Get a scalar real.

   use extendr_api::*;
   extendr_engine::start_r();
   let robj1 = Robj::from(1);
   let robj2 = Robj::from(1.);
   let robj3 = Robj::from(NA_REAL);
   assert_eq!(robj1.as_real(), None);
   assert_eq!(robj2.as_real(), Some(1.));
   assert_eq!(robj3.as_real(), None);

pub fn as_bool(&self) -> Option<bool>

Get a scalar rust boolean.

   use extendr_api::*;
   extendr_engine::start_r();
   let robj1 = Robj::from(TRUE);
   let robj2 = Robj::from(1.);
   let robj3 = Robj::from(NA_LOGICAL);
   assert_eq!(robj1.as_bool(), Some(true));
   assert_eq!(robj2.as_bool(), None);
   assert_eq!(robj3.as_bool(), None);

pub fn as_logical(&self) -> Option<Bool>

Get a scalar boolean as a tri-boolean Bool value.

   use extendr_api::*;
   extendr_engine::start_r();
   let robj1 = Robj::from(TRUE);
   let robj2 = Robj::from([TRUE, FALSE]);
   let robj3 = Robj::from(NA_LOGICAL);
   assert_eq!(robj1.as_logical(), Some(TRUE));
   assert_eq!(robj2.as_logical(), None);
   assert_eq!(robj3.as_logical(), Some(NA_LOGICAL));

pub fn eval(&self) -> Result<Robj, AnyError>

Evaluate the expression in R and return an error or an R object.

   use extendr_api::*;
   extendr_engine::start_r();

   let add = lang!("+", 1, 2);
   assert_eq!(add.eval().unwrap(), r!(3));

pub fn eval_blind(&self) -> Robj

Evaluate the expression and return NULL or an R object.

   use extendr_api::*;
   extendr_engine::start_r();
   let bad = lang!("imnotavalidfunctioninR", 1, 2);
   assert_eq!(bad.eval_blind(), r!(NULL));

pub fn parse(code: &str) -> Result<Robj, AnyError>

Parse a string into an R executable object

   use extendr_api::*;
   extendr_engine::start_r();
   let expr = Robj::parse("1 + 2").unwrap();
   assert!(expr.is_expression());

pub fn eval_string(code: &str) -> Result<Robj, AnyError>

Parse a string into an R executable object and run it.

   use extendr_api::*;
   extendr_engine::start_r();
   let res = Robj::eval_string("1 + 2").unwrap();
   assert_eq!(res, r!(3.));

pub fn is_owned(&self) -> bool

Return true if the object is owned by this wrapper. If so, it will be released when the wrapper drops.

   use extendr_api::*;
   extendr_engine::start_r();
   let owned = r!(1);      // Allocated vector.
   let borrowed = r!(());  // R_NilValue
   assert_eq!(owned.is_owned(), true);
   assert_eq!(borrowed.is_owned(), false);

impl Robj

These are helper functions which give access to common properties of R objects.

pub fn get_attrib<'a, N>(&self, name: N) -> Option<Robj> where
Self: 'a,
Robj: From<N> + 'a,

Get a specific attribute as a borrowed robj if it exists.

   use extendr_api::*;
   extendr_engine::start_r();

   let mut robj = r!("hello");
   robj.set_attrib(Symbol("xyz"), 1);
   assert_eq!(robj.get_attrib(Symbol("xyz")), Some(r!(1)));

pub fn set_attrib<N, V>(&mut self, name: N, value: V) -> Robj where
Robj: From<N>,
Robj: From<V>,

Set a specific attribute and return the value.

   use extendr_api::*;
   extendr_engine::start_r();

   let mut robj = r!("hello");
   let value = robj.set_attrib(Symbol("xyz"), 1);
   assert_eq!(robj.get_attrib(Symbol("xyz")), Some(value));

pub fn names(&self) -> Option<StrIter>

Get the names attribute as a string iterator if one exists.

   use extendr_api::*;
   extendr_engine::start_r();

   let list = list!(a = 1, b = 2, c = 3);
   let names : Vec<_> = list.names().unwrap().collect();
   assert_eq!(names, vec!["a", "b", "c"]);

pub fn dim<'a>(&self) -> Option<IntegerIter<'a>> where
Self: 'a,
Robj: 'a,

Get the dim attribute as an integer iterator if one exists.

   use extendr_api::*;
   extendr_engine::start_r();

   let array = R!(array(data = c(1, 2, 3, 4), dim = c(2, 2), dimnames = list(c("x", "y"), c("a","b")))).unwrap();
   let dim : Vec<_> = array.dim().unwrap().cloned().collect();
   assert_eq!(dim, vec![2, 2]);

pub fn dimnames(&self) -> Option<VecIter>

Get the dimnames attribute as a list iterator if one exists.

   use extendr_api::*;
   extendr_engine::start_r();

   let array = R!(array(data = c(1, 2, 3, 4), dim = c(2, 2), dimnames = list(c("x", "y"), c("a","b")))).unwrap();
   let names : Vec<_> = array.dimnames().unwrap().collect();
   assert_eq!(names, vec![r!(["x", "y"]), r!(["a", "b"])]);

pub fn named_list_iter(&self) -> Option<NamedListIter>

Return an iterator over names and values of a list if they exist.

   use extendr_api::*;
   extendr_engine::start_r();

   let list = list!(a = 1, b = 2, c = 3);
   let names_and_values : Vec<_> = list.named_list_iter().unwrap().collect();
   assert_eq!(names_and_values, vec![("a", r!(1)), ("b", r!(2)), ("c", r!(3))]);

pub fn class(&self) -> Option<StrIter>

Get the class attribute as a string iterator if one exists.

   use extendr_api::*;
   extendr_engine::start_r();

   let formula = R!(y ~ A * x + b).unwrap();
   let class : Vec<_> = formula.class().unwrap().collect();
   assert_eq!(class, ["formula"]);

pub fn inherits(&self, classname: &str) -> bool

Return true if this class inherits this class.

   use extendr_api::*;
   extendr_engine::start_r();

   let formula = R!(y ~ A * x + b).unwrap();
   assert_eq!(formula.inherits("formula"), true);

pub fn levels(&self) -> Option<StrIter>

Get the levels attribute as a string iterator if one exists.

   use extendr_api::*;
   extendr_engine::start_r();

   let factor = factor!(vec!["abcd", "def", "fg", "fg"]);
   let levels : Vec<_> = factor.levels().unwrap().collect();
   assert_eq!(levels, vec!["abcd", "def", "fg"]);