Struct extendr_api::graphics::Context

pub struct Context { /* private fields */ }

Implementations

impl Context

pub fn from_device(dev: &Device, unit: Unit) -> Self

pub fn color(&mut self, col: Color) -> &mut Self

Set the line or text color of a primitive.

pub fn fill(&mut self, fill: Color) -> &mut Self

Set the fill color of a primitive.

pub fn gamma(&mut self, gamma: f64) -> &mut Self

Set the gamma of the device. out_color = in_color ** gamma

pub fn line_width(&mut self, lwd: f64) -> &mut Self

Set the width of the line in chosen units.

pub fn line_type(&mut self, lty: LineType) -> &mut Self

Set the type of the line.

Blank    => <invisible>
Solid    => ------
Dashed   => - - - -
Dotted   => . . . .
DotDash  => . - . -
LongDash => --  --
TwoDash  => . . - -

pub fn line_end(&mut self, lend: LineEnd) -> &mut Self

Set the line end type.

  LineEnd::RoundCap
  LineEnd::ButtCap  
  LineEnd::SquareCap

pub fn line_join(&mut self, ljoin: LineJoin) -> &mut Self

Set the line join type.

  LineJoin::RoundJoin
  LineJoin::MitreJoin
  LineJoin::BevelJoin

pub fn point_size(&mut self, ps: f64) -> &mut Self

pub fn line_mitre(&mut self, lmitre: f64) -> &mut Self

Set the line miter limit - the point where the line becomes a bevel join.

pub fn line_height(&mut self, lineheight: f64) -> &mut Self

Set the line height for text.

pub fn font_face(&mut self, fontface: FontFace) -> &mut Self

Set the font face.

  FontFace::PlainFont
  FontFace::BoldFont
  FontFace::ItalicFont
  FontFace::BoldItalicFont
  FontFace::SymbolFont

pub fn font_family(&mut self, fontfamily: &str) -> &mut Self

pub fn transform(
&mut self,
xscale: (f64, f64),
yscale: (f64, f64),
offset: (f64, f64)
) -> &mut Self

Set the transform as a 3x2 matrix.