extendr_api::graphics

Type Alias GEDevDesc

pub type GEDevDesc = _GEDevDesc;

Aliased Type

#[repr(C)]
pub struct GEDevDesc {
    pub dev: *mut _DevDesc,
    pub displayListOn: Rboolean,
    pub displayList: *mut SEXPREC,
    pub DLlastElt: *mut SEXPREC,
    pub savedSnapshot: *mut SEXPREC,
    pub dirty: Rboolean,
    pub recordGraphics: Rboolean,
    pub gesd: [*mut GESystemDesc; 24],
    pub ask: Rboolean,
    pub appending: Rboolean,
}

Fields
dev: *mut _DevDesc

Stuff that the devices can see (and modify). All detailed in GraphicsDevice.h

displayListOn: Rboolean

toggle for display list status

displayList: *mut SEXPREC

display list

DLlastElt: *mut SEXPREC

A pointer to the end of the display list to avoid traversing pairlists

savedSnapshot: *mut SEXPREC

The last element of the display list just prior to when the display list was last initialised

dirty: Rboolean

Has the device received any output?

recordGraphics: Rboolean

Should a graphics call be stored on the display list? Set to FALSE by do_recordGraphics, do_dotcallgr, and do_Externalgr so that nested calls are not recorded on the display list

gesd: [*mut GESystemDesc; 24]

Stuff about the device that only graphics systems see. The graphics engine has no idea what is in here. Used by graphics systems to store system state per device.

ask: Rboolean

per-device setting for ‘ask’ (use NewFrameConfirm)

appending: Rboolean

Is a device appending a path ?