Macro extendr_api::test

macro_rules! test {
    () => { ... };
    ($($rest: tt)*) => { ... };
}

Macro for running tests.

This starts up the underlying extendr_engine so that interactions with R will work. Additionally, this allows us to use ? in example code instead of unwrap().

Note: This macro is meant to be used in test code (annotated with #[cfg(test)]) or in doc strings. If it is used in library code that gets incorporated into an R package, R CMD check will complain about non-API calls.