extendr_api
Module optional

A set of optional features and third-party crate integrations, usually hidden behind feature gates.

Modules

either

Enables support for the either crate, to allow accepting and returning Either<L, R> values if both L and R are convertible to/from Robj.

faer 🔒

ndarray

Defines conversions between R objects and the ndarray crate, which offers native Rust array types and numerical computation routines.