extendr_api::prelude::col
Trait ColBatch

pub trait ColBatch<E>: As2D<E>
where
    E: Conjugate,
{
    type Owned: ColBatchMut<<E as Conjugate>::Canonical>;

    // Required methods
    fn new_owned_zeros(nrows: usize, ncols: usize) -> Self::Owned;
    fn new_owned_copied(src: &Self) -> Self::Owned;
    fn resize_owned(owned: &mut Self::Owned, nrows: usize, ncols: usize);
}

Type that can be interpreted as a batch of column vectors. Can be a single column or a matrix.

Required Associated Types

type Owned: ColBatchMut<<E as Conjugate>::Canonical>

Corresponding owning type.

Required Methods

fn new_owned_zeros(nrows: usize, ncols: usize) -> Self::Owned

Constructor of the owned type that initializes the values to zero.

fn new_owned_copied(src: &Self) -> Self::Owned

Constructor of the owned type that copies the values.

fn resize_owned(owned: &mut Self::Owned, nrows: usize, ncols: usize)

Resize an owned column or matrix.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.