Module extendr_api::prelude::mat

Matrix type.

Structs

Mat

Heap allocated resizable matrix, similar to a 2D Vec.

MatMut

Mutable view over a matrix, similar to a mutable reference to a 2D strided slice.

MatRef

Immutable view over a matrix, similar to an immutable reference to a 2D strided slice.

Traits

As2D

Trait for types that can be converted to a 2D matrix view.

As2DMut

Trait for types that can be converted to a mutable 2D matrix view.

AsMatMut

Trait for types that can be converted to a mutable matrix view.

AsMatRef

Trait for types that can be converted to a matrix view.

MatIndex

Represents a type that can be used to slice a matrix, such as an index or a range of indices.

Functions

from_column_major_slice

Creates a MatRef from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a column-major format, so that the first chunk of nrows values from the slices goes in the first column of the matrix, the second chunk of nrows values goes in the second column, and so on.

from_column_major_slice_mut

Creates a MatMut from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a column-major format, so that the first chunk of nrows values from the slices goes in the first column of the matrix, the second chunk of nrows values goes in the second column, and so on.

from_column_major_slice_with_stride

Creates a MatRef from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a column-major format, where the beginnings of two consecutive columns are separated by col_stride elements.

from_column_major_slice_with_stride_mut

Creates a MatMut from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a column-major format, where the beginnings of two consecutive columns are separated by col_stride elements.

from_raw_parts⚠

Creates a MatRef from pointers to the matrix data, dimensions, and strides.

from_raw_parts_mut⚠

Creates a MatMut from pointers to the matrix data, dimensions, and strides.

from_row_major_slice

Creates a MatRef from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a row-major format, so that the first chunk of ncols values from the slices goes in the first column of the matrix, the second chunk of ncols values goes in the second column, and so on.

from_row_major_slice_mut

Creates a MatMut from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a row-major format, so that the first chunk of ncols values from the slices goes in the first column of the matrix, the second chunk of ncols values goes in the second column, and so on.

from_row_major_slice_with_stride

Creates a MatRef from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a row-major format, where the beginnings of two consecutive rows are separated by row_stride elements.

from_row_major_slice_with_stride_mut

Creates a MatMut from slice views over the matrix data, and the matrix dimensions. The data is interpreted in a row-major format, where the beginnings of two consecutive rows are separated by row_stride elements.