extendr_api::prelude
Module modules

👎Deprecated

Re-exports.

Modules

cholesky

Low level implementation of the various Cholesky-like decompositions.

coreDeprecated

Emulation layer for faer_core

evd

Low level implementation of the eigenvalue decomposition of a square diagonalizable matrix.

lu

Low level implementation of the LU decompositions.

qr

This crate provides utilities for computing and manipulating the QR factorization with and without pivoting. The QR factorization decomposes a matrix into a product of a unitary matrix $Q$ (represented using block Householder sequences), and an upper trapezoidal matrix $R$, such that their product is equal to the original matrix (or a column permutation of it in the case where column pivoting is used).

svd

Low level implementation of the SVD of a matrix.