extendr_api::prelude::modules::cholesky::bunch_kaufman::compute
Function cholesky_in_place

pub fn cholesky_in_place<'out, I, E>(
    matrix: MatMut<'_, E>,
    subdiag: ColMut<'_, E>,
    regularization: BunchKaufmanRegularization<'_, E>,
    perm: &'out mut [I],
    perm_inv: &'out mut [I],
    parallelism: Parallelism<'_>,
    stack: PodStack<'_>,
    params: BunchKaufmanParams,
) -> (BunchKaufmanInfo, PermRef<'out, I>)
where
    I: Index,
    E: ComplexField,

Computes the Cholesky factorization with Bunch-Kaufman pivoting of the input matrix and stores the factorization in matrix and subdiag.

The diagonal of the block diagonal matrix is stored on the diagonal of matrix, while the subdiagonal elements of the blocks are stored in subdiag.

Panics

Panics if the input matrix is not square.

This can also panic if the provided memory in stack is insufficient (see cholesky_in_place_req).