extendr_api::prelude::modules::cholesky::bunch_kaufman
Module solve

Solving a linear system using the decomposition.

Functions

solve_in_place_req

Computes the size and alignment of required workspace for solving a linear system defined by a matrix in place, given its Bunch-Kaufman decomposition.

solve_in_place_with_conj

Given the Bunch-Kaufman factors of a matrix $A$ and a matrix $B$ stored in rhs, this function computes the solution of the linear system: $$\text{Op}_A(A)X = B.$$