Module extendr_api::prelude::modules::cholesky::ldlt_diagonal::compute

Computing the decomposition.

Structs

LdltDiagParams

LDLT factorization tuning parameters.

LdltInfo

Info about the result of the LDLT factorization.

LdltRegularization

Dynamic LDLT regularization. Values below epsilon in absolute value, or with the wrong sign are set to delta with their corrected sign.

Functions

raw_cholesky_in_place

Computes the Cholesky factors $L$ and $D$ of the input matrix such that $L$ is strictly lower triangular, $D$ is real-valued diagonal, and $$LDL^H = A.$$

raw_cholesky_in_place_req

Computes the size and alignment of required workspace for performing a Cholesky decomposition.