Module extendr_api::prelude::modules::cholesky::llt

The Cholesky decomposition of a Hermitian positive definite matrix $A$ is such that: $$A = LL^H,$$ where $L$ is a lower triangular matrix.

Modules

compute

Computing the decomposition.

inverse

Reconstructing the inverse of the original matrix from the decomposition.

reconstruct

Reconstructing the original matrix from the decomposition.

solve

Solving a linear system using the decomposition.

update

Updating the decomposition.

Structs

CholeskyError

This error signifies that the LLT decomposition could not be computed due to the matrix not being numerically positive definite.