Function extendr_api::prelude::modules::cholesky::llt::solve::solve_with_conj

pub fn solve_with_conj<E>(
    dst: MatMut<'_, E>,
    cholesky_factor: MatRef<'_, E>,
    conj_lhs: Conj,
    rhs: MatRef<'_, E>,
    parallelism: Parallelism,
    stack: PodStack<'_>,
)
where
    E: ComplexField,

Given the Cholesky factor of a matrix $A$ and a matrix $B$ stored in rhs, this function computes the solution of the linear system: $$\text{Op}_A(A)X = B.$$

$\text{Op}_A$ is either the identity or the conjugation depending on the value of conj_lhs.

The solution of the linear system is stored in dst.

Panics

Panics if any of these conditions is violated:

cholesky_factor must be square of dimension n.

rhs must have n rows.

dst must have n rows, and the same number of columns as rhs.

This can also panic if the provided memory in stack is insufficient (see solve_req).