extendr_api::prelude::modules::cholesky::llt
Module update

Updating the decomposition.

Functions

delete_rows_and_cols_clobber

Deletes r rows and columns at the provided indices from the Cholesky factor.

delete_rows_and_cols_clobber_req

Computes the size and alignment of required workspace for deleting the rows and columns from a matrix, given its Cholesky decomposition.

insert_rows_and_cols_clobber

Inserts r rows and columns at the provided index in the Cholesky factor.

insert_rows_and_cols_clobber_req

Computes the size and alignment of the required workspace for inserting the rows and columns at the index in the Cholesky factor..

rank_r_update_clobber

Performs a rank-r update in place, while clobbering the inputs.