Module extendr_api::prelude::modules::core

๐Deprecated

Emulation layer for faer_core

Modules

householder

Block Householder transformations.

inverse

Triangular matrix inversion.

mul

Matrix multiplication.

zip

Implementation of zipped! structures.

Macros

col

Creates a col::Col containing the arguments.

mat

Creates a Mat containing the arguments.

row

Creates a row::Row containing the arguments.

unzipped

Used to undo the zipping by the zipped! macro.

zipped

Zips together matrix of the same size, so that coefficient-wise operations can be performed on their elements.

Structs

Col

Heap allocated resizable column vector.

ColMut

Mutable view over a column vector, similar to a mutable reference to a strided slice.

ColRef

Immutable view over a column vector, similar to an immutable reference to a strided slice.

IdentityGroup

Mat

Heap allocated resizable matrix, similar to a 2D Vec.

MatMut

Mutable view over a matrix, similar to a mutable reference to a 2D strided slice.

MatRef

Immutable view over a matrix, similar to an immutable reference to a 2D strided slice.

Row

Heap allocated resizable row vector.

RowMut

Mutable view over a row vector, similar to a mutable reference to a strided slice.

RowRef

Immutable view over a row vector, similar to an immutable reference to a strided slice.

Scale

Factor for matrix-scalar multiplication.

Enums

Conj

Whether a matrix should be implicitly conjugated when read or not.

FaerError

Errors that can occur in sparse algorithms.

Parallelism

Parallelism strategy that can be passed to most of the routines in the library.

Side

Specifies whether the triangular lower or upper part of a matrix should be accessed.

Traits

As2D

Trait for types that can be converted to a 2D matrix view.

As2DMut

Trait for types that can be converted to a mutable 2D matrix view.

AsColMut

Trait for types that can be converted to a mutable column view.

AsColRef

Trait for types that can be converted to a column view.

AsMatMut

Trait for types that can be converted to a mutable matrix view.

AsMatRef

Trait for types that can be converted to a matrix view.

AsRowMut

Trait for types that can be converted to a mutable row view.

AsRowRef

Trait for types that can be converted to a row view.

ComplexField

Unstable trait containing the operations that a number type needs to implement.

Conjugate

Trait for types that may be implicitly conjugated.

Entity

Unstable core trait for describing how a scalar value may be split up into individual component.

RealField

Unstable trait containing the operations that a real number type needs to implement.

SimpleEntity

Functions

disable_global_parallelism

Causes functions that access global parallelism settings to panic.

get_global_parallelism

Gets the global parallelism settings.

kron

Kronecker product of two matrices.

scale

Returns a factor for matrix-scalar multiplication.

set_global_parallelism

Sets the global parallelism settings.

temp_mat_constant

Creates a temporary matrix of constant values, from the given memory stack.

temp_mat_req

Returns the stack requirements for creating a temporary matrix with the given dimensions.

temp_mat_uninit

Creates a temporary matrix of untouched values, from the given memory stack.

temp_mat_zeroed

Creates a temporary matrix of zero values, from the given memory stack.