extendr_api::prelude::modules::core::householder
Function apply_block_householder_transpose_on_the_right_in_place_with_conj

pub fn apply_block_householder_transpose_on_the_right_in_place_with_conj<E>(
    householder_basis: MatRef<'_, E>,
    householder_factor: MatRef<'_, E>,
    conj_rhs: Conj,
    matrix: MatMut<'_, E>,
    parallelism: Parallelism<'_>,
    stack: PodStack<'_>,
)
where
    E: ComplexField,

Computes the product of the matrix, multiplied by the transpose of the given block Householder transformation, and stores the result in matrix.