Module extendr_api::prelude::modules::core::mul::triangular

Triangular matrix multiplication module, where some of the operands are treated as triangular matrices.

Enums

BlockStructure

Describes the parts of the matrix that must be accessed.

Functions

matmul

Computes the matrix product [alpha * acc] + beta * lhs * rhs and stores the result in acc.

matmul_with_conj

Computes the matrix product [alpha * acc] + beta * lhs * rhs (while optionally conjugating either or both of the input matrices) and stores the result in acc.