Struct extendr_api::prelude::modules::core::ColRef

#[repr(C)]
pub struct ColRef<'a, E>
where
    E: Entity,
{
    pub(super) inner: VecImpl<E>,
    pub(super) __marker: PhantomData<&'a E>,
}

Immutable view over a column vector, similar to an immutable reference to a strided slice.

Note

Unlike a slice, the data pointed to by ColRef<'_, E> is allowed to be partially or fully uninitialized under certain conditions. In this case, care must be taken to not perform any operations that read the uninitialized values, or form references to them, either directly through ColRef::read, or indirectly through any of the numerical library routines, unless it is explicitly permitted.

Fields
inner: VecImpl<E>__marker: PhantomData<&'a E>
Implementations

impl<'a, E> ColRef<'a, E>
where E: Entity,

pub fn nrows(&self) -> usize

Returns the number of rows of the column.

pub fn ncols(&self) -> usize

Returns the number of columns of the column. This is always equal to 1.

pub fn as_ptr( self ) -> <<E as Entity>::Group as ForType>::FaerOf<*const <E as Entity>::Unit>

Returns pointers to the matrix data.

pub fn row_stride(&self) -> isize

Returns the row stride of the matrix, specified in number of elements, not in bytes.

pub fn as_2d(self) -> MatRef<'a, E>

Returns self as a matrix view.

pub fn ptr_at( self, row: usize ) -> <<E as Entity>::Group as ForType>::FaerOf<*const <E as Entity>::Unit>

Returns raw pointers to the element at the given index.

pub unsafe fn ptr_inbounds_at( self, row: usize ) -> <<E as Entity>::Group as ForType>::FaerOf<*const <E as Entity>::Unit>

Returns raw pointers to the element at the given index, assuming the provided index is within the size of the vector.

Safety

The behavior is undefined if any of the following conditions are violated:

row < self.nrows().

pub unsafe fn split_at_unchecked( self, row: usize ) -> (ColRef<'a, E>, ColRef<'a, E>)

Splits the column vector at the given index into two parts and returns an array of each subvector, in the following order:

top.

bottom.

Safety

The behavior is undefined if any of the following conditions are violated:

row <= self.nrows().

pub unsafe fn split_at(self, row: usize) -> (ColRef<'a, E>, ColRef<'a, E>)

Splits the column vector at the given index into two parts and returns an array of each subvector, in the following order:

top.

bottom.

Panics

The function panics if any of the following conditions are violated:

row <= self.nrows().

pub unsafe fn get_unchecked<RowRange>( self, row: RowRange ) -> <ColRef<'a, E> as ColIndex<RowRange>>::Target
where ColRef<'a, E>: ColIndex<RowRange>,

Returns references to the element at the given index, or subvector if row is a range.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

Safety

The behavior is undefined if any of the following conditions are violated:

row must be contained in [0, self.nrows()).

pub fn get<RowRange>( self, row: RowRange ) -> <ColRef<'a, E> as ColIndex<RowRange>>::Target
where ColRef<'a, E>: ColIndex<RowRange>,

Returns references to the element at the given index, or subvector if row is a range, with bound checks.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

Panics

The function panics if any of the following conditions are violated:

row must be contained in [0, self.nrows()).

pub unsafe fn read_unchecked(&self, row: usize) -> E

Reads the value of the element at the given index.

Safety

The behavior is undefined if any of the following conditions are violated:

row < self.nrows().

pub fn read(&self, row: usize) -> E

Reads the value of the element at the given index, with bound checks.

Panics

The function panics if any of the following conditions are violated:

row < self.nrows().

pub fn transpose(self) -> RowRef<'a, E>

Returns a view over the transpose of self.

pub fn conjugate(self) -> ColRef<'a, <E as Conjugate>::Conj>
where E: Conjugate,

Returns a view over the conjugate of self.

pub fn adjoint(self) -> RowRef<'a, <E as Conjugate>::Conj>
where E: Conjugate,

Returns a view over the conjugate transpose of self.

pub fn canonicalize(self) -> (ColRef<'a, <E as Conjugate>::Canonical>, Conj)
where E: Conjugate,

Returns a view over the canonical representation of self, as well as a flag declaring whether self is implicitly conjugated or not.

pub fn reverse_rows(self) -> ColRef<'a, E>

Returns a view over the self, with the rows in reversed order.

pub unsafe fn subrows_unchecked( self, row_start: usize, nrows: usize ) -> ColRef<'a, E>

Returns a view over the subvector starting at row row_start, and with number of rows nrows.

Safety

The behavior is undefined if any of the following conditions are violated:

row_start <= self.nrows().

nrows <= self.nrows() - row_start.

pub fn subrows(self, row_start: usize, nrows: usize) -> ColRef<'a, E>

Returns a view over the subvector starting at row row_start, and with number of rows nrows.

Safety

The behavior is undefined if any of the following conditions are violated:

row_start <= self.nrows().

nrows <= self.nrows() - row_start.

pub fn column_vector_as_diagonal(self) -> DiagRef<'a, E>

Given a matrix with a single column, returns an object that interprets the column as a diagonal matrix, whoes diagonal elements are values in the column.

pub fn to_owned(&self) -> Col<<E as Conjugate>::Canonical>
where E: Conjugate,

Returns an owning Col of the data.

pub fn has_nan(&self) -> bool
where E: ComplexField,

Returns true if any of the elements is NaN, otherwise returns false.

pub fn is_all_finite(&self) -> bool
where E: ComplexField,

Returns true if all of the elements are finite, otherwise returns false.

pub fn norm_max(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the maximum norm of self.

pub fn norm_l1(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the L1 norm of self.

pub fn norm_l2(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the L2 norm of self.

pub fn squared_norm_l2(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the squared L2 norm of self.

pub fn sum(&self) -> E
where E: ComplexField,

Returns the sum of self.

pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
where E: ComplexField,

Kroneckor product of self and rhs.

This is an allocating operation; see faer::linalg::kron for the allocation-free version or more info in general.

pub fn try_as_slice( self ) -> Option<<<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>>

Returns the column as a contiguous slice if its row stride is equal to 1.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

pub fn as_ref(&self) -> ColRef<'_, E>

Returns a view over the matrix.