extendr_api::prelude::modules::core
Trait As2D

pub trait As2D<E>
where
    E: Entity,
{
    // Required method
    fn as_2d_ref(&self) -> MatRef<'_, E>;
}

Trait for types that can be converted to a 2D matrix view.

This trait is implemented for any type that can be represented as a 2D matrix view, like Mat, Row, Col, and their respective references and mutable references. For a trait specific to the matrix family, see AsMatRef or AsMatMut.

Required Methods

fn as_2d_ref(&self) -> MatRef<'_, E>

Convert to a 2D matrix view.