Struct extendr_api::prelude::modules::core::zip::ZipEq

pub struct ZipEq<Rows, Cols, Head, Tail>(Head, Tail)
where
    Head: MatShape<Rows = Rows, Cols = Cols>,
    Tail: MatShape<Rows = Rows, Cols = Cols>;

Zipped matrix views.

Tuple Fields
0: Head1: Tail
Implementations

impl<Rows, Cols, Head, Tail> ZipEq<Rows, Cols, Head, Tail>
where Rows: Copy + Eq, Cols: Copy + Eq, Head: MatShape<Rows = Rows, Cols = Cols>, Tail: MatShape<Rows = Rows, Cols = Cols>,

pub fn new(head: Head, tail: Tail) -> ZipEq<Rows, Cols, Head, Tail>

Creates a zipped matrix, after asserting that the dimensions match.

pub fn new_unchecked(head: Head, tail: Tail) -> ZipEq<Rows, Cols, Head, Tail>

Creates a zipped matrix, assuming that the dimensions match.

impl<Head, Tail> ZipEq<(), usize, Head, Tail>
where Head: for<'a> MatIndex<'a, Rows = (), Cols = usize, Index = ((), usize), LayoutTransform = VecLayoutTransform>, Tail: for<'a> MatIndex<'a, Rows = (), Cols = usize, Index = ((), usize), LayoutTransform = VecLayoutTransform>,

pub fn for_each( self, f: impl for<'a> FnMut(<ZipEq<(), usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of self.

pub fn for_each_with_index( self, f: impl for<'a> FnMut(usize, <ZipEq<(), usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of self, while passing in the index of the current element.

pub fn map<E>( self, f: impl for<'a> FnMut(<ZipEq<(), usize, Head, Tail> as MatIndex<'a>>::Item) -> E ) -> Row<E>
where E: Entity,

Applies f to each element of self and collect its result into a new row.

impl<Head, Tail> ZipEq<usize, (), Head, Tail>
where Head: for<'a> MatIndex<'a, Rows = usize, Cols = (), Index = (usize, ()), LayoutTransform = VecLayoutTransform>, Tail: for<'a> MatIndex<'a, Rows = usize, Cols = (), Index = (usize, ()), LayoutTransform = VecLayoutTransform>,

pub fn for_each( self, f: impl for<'a> FnMut(<ZipEq<usize, (), Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of self.

pub fn for_each_with_index( self, f: impl for<'a> FnMut(usize, <ZipEq<usize, (), Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of self, while passing in the index of the current element.

pub fn map<E>( self, f: impl for<'a> FnMut(<ZipEq<usize, (), Head, Tail> as MatIndex<'a>>::Item) -> E ) -> Col<E>
where E: Entity,

Applies f to each element of self and collect its result into a new column.

impl<Head, Tail> ZipEq<usize, usize, Head, Tail>
where Head: for<'a> MatIndex<'a, Rows = usize, Cols = usize, Index = (usize, usize), LayoutTransform = MatLayoutTransform>, Tail: for<'a> MatIndex<'a, Rows = usize, Cols = usize, Index = (usize, usize), LayoutTransform = MatLayoutTransform>,

pub fn for_each( self, f: impl for<'a> FnMut(<ZipEq<usize, usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of self.

pub fn for_each_with_index( self, f: impl for<'a> FnMut(usize, usize, <ZipEq<usize, usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of self, while passing the indices of the position of the current element.

pub fn for_each_triangular_lower_with_index( self, diag: Diag, f: impl for<'a> FnMut(usize, usize, <ZipEq<usize, usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of the lower triangular half of self, while passing the indices of the position of the current element.

diag specifies whether the diagonal should be included or excluded.

pub fn for_each_triangular_upper_with_index( self, diag: Diag, f: impl for<'a> FnMut(usize, usize, <ZipEq<usize, usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of the upper triangular half of self, while passing the indices of the position of the current element.

diag specifies whether the diagonal should be included or excluded.

pub fn for_each_triangular_lower( self, diag: Diag, f: impl for<'a> FnMut(<ZipEq<usize, usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of the lower triangular half of self.

diag specifies whether the diagonal should be included or excluded.

pub fn for_each_triangular_upper( self, diag: Diag, f: impl for<'a> FnMut(<ZipEq<usize, usize, Head, Tail> as MatIndex<'a>>::Item) )

Applies f to each element of the upper triangular half of self.

diag specifies whether the diagonal should be included or excluded.

pub fn map<E>( self, f: impl for<'a> FnMut(<ZipEq<usize, usize, Head, Tail> as MatIndex<'a>>::Item) -> E ) -> Mat<E>
where E: Entity,

Applies f to each element of self and collect its result into a new matrix.