Module extendr_api::prelude::modules::evd

Low level implementation of the eigenvalue decomposition of a square diagonalizable matrix.

The eigenvalue decomposition of a square matrix $M$ of shape $(n, n)$ is a decomposition into two components $U$, $S$:

$U$ has shape $(n, n)$ and is invertible,

$S$ has shape $(n, n)$ and is a diagonal matrix,

and finally:

$$M = U S U^{-1}.$$

If $M$ is hermitian, then $U$ can be made unitary ($U^{-1} = U^H$), and $S$ is real valued.

Structs

EvdParams

HermitianEvdParams

Hermitian eigendecomposition tuning parameters.

Enums

ComputeVectors

Indicates whether the eigenvectors are fully computed, partially computed, or skipped.

Functions

compute_evd_complex

Computes the eigenvalue decomposition of a square complex matrix.

compute_evd_complex_custom_epsilon

See compute_evd_complex.

compute_evd_real

Computes the eigenvalue decomposition of a square real matrix.

compute_evd_real_custom_epsilon

See compute_evd_real.

compute_evd_req

Computes the size and alignment of required workspace for performing an eigenvalue decomposition. The eigenvectors may be optionally computed.

compute_hermitian_evd

Computes the eigenvalue decomposition of a square hermitian matrix. Only the lower triangular half of the matrix is accessed.

compute_hermitian_evd_custom_epsilon

See compute_hermitian_evd.

compute_hermitian_evd_req

Computes the size and alignment of required workspace for performing a hermitian eigenvalue decomposition. The eigenvectors may be optionally computed.