extendr_api::prelude::modules::evd
Function compute_hermitian_pseudoinverse_custom_epsilon

pub fn compute_hermitian_pseudoinverse_custom_epsilon<E>(
    pinv: MatMut<'_, E>,
    s: ColRef<'_, E>,
    u: MatRef<'_, E>,
    atol: Option<<E as ComplexField>::Real>,
    rtol: Option<<E as ComplexField>::Real>,
    parallelism: Parallelism<'_>,
    stack: PodStack<'_>,
)
where
    E: ComplexField,

See compute_hermitian_pseudoinverse.

This function takes an additional atol and rtol parameters. atol represents the absolute threshold term, if None provided value is 0, and rtol represents the relative threshold term, if None provided value is N * eps where eps is the machine precision value of the E::Real

Panics

Panics if atol or rtol equal or less than zero