extendr_api::prelude::modules::lu::partial_pivoting
Module compute

Computing the decomposition.

Structs

PartialPivLuComputeParams

LUfactorization tuning parameters.

PartialPivLuInfo

Information about the resulting LU factorization.

Functions

lu_in_place

Computes the LU decomposition of the given matrix with partial pivoting, replacing the matrix with its factors in place.

lu_in_place_req

Computes the size and alignment of required workspace for performing an LU decomposition with partial pivoting.