Function extendr_api::prelude::modules::lu::partial_pivoting::solve::solve_transpose

pub fn solve_transpose<I, E>(
    dst: MatMut<'_, E>,
    lu_factors: MatRef<'_, E>,
    conj_lhs: Conj,
    row_perm: PermRef<'_, I>,
    rhs: MatRef<'_, E>,
    parallelism: Parallelism,
    stack: PodStack<'_>
)
where
    I: Index,
    E: ComplexField,

Given the LU factors of a matrix $A$ and a matrix $B$ stored in rhs, this function computes the solution of the linear system: $$\text{Op}_A(A)^\top X = B.$$

$\text{Op}_A$ is either the identity or the conjugation depending on the value of conj_lhs.

The solution of the linear system is stored in dst.

Panics

Panics if lu_factors is not a square matrix.

Panics if row_perm doesn’t have the same dimension as lu_factors.

Panics if rhs doesn’t have the same number of rows as the dimension of lu_factors.

Panics if rhs and dst don’t have the same shape.

Panics if the provided memory in stack is insufficient (see solve_transpose_req).