extendr_api::prelude::modules::qr::col_pivoting
Module compute

Computing the decomposition.

Structs

ColPivQrComputeParams

QR factorization tuning parameters.

ColPivQrInfo

Information about the resulting QR factorization.

Functions

qr_in_place

Computes the QR decomposition with pivoting of a rectangular matrix $A$, into a unitary matrix $Q$, represented as a block Householder sequence, and an upper trapezoidal matrix $R$, such that $$AP^\top = QR.$$

qr_in_place_req

Computes the size and alignment of required workspace for performing a QR decomposition with column pivoting.

recommended_blocksize

The recommended block size to use for a QR decomposition of a matrix with the given shape.