extendr_api::prelude::modules::qr::no_pivoting::inverse
Function invert_in_place_req

pub fn invert_in_place_req<E>(
    qr_nrows: usize,
    qr_ncols: usize,
    blocksize: usize,
    parallelism: Parallelism<'_>,
) -> Result<StackReq, SizeOverflow>
where
    E: Entity,

Computes the size and alignment of required workspace for computing the inverse of a matrix in place, given its QR decomposition.