Module extendr_api::prelude::modules::qr::no_pivoting::reconstruct

Reconstructing the original matrix from the decomposition.

Functions

reconstruct

Computes the reconstructed matrix, given its QR decomposition, and stores the result in dst.

reconstruct_in_place

Computes the reconstructed matrix, given its QR decomposition, and stores the result in qr_factors.

reconstruct_in_place_req

Computes the size and alignment of required workspace for reconstructing a matrix in place, given its QR decomposition.

reconstruct_req

Computes the size and alignment of required workspace for reconstructing a matrix out of place, given its QR decomposition.