Function extendr_api::prelude::modules::qr::no_pivoting::reconstruct::reconstruct

pub fn reconstruct<E>(
    dst: MatMut<'_, E>,
    qr_factors: MatRef<'_, E>,
    householder_factor: MatRef<'_, E>,
    parallelism: Parallelism,
    stack: PodStack<'_>,
)
where
    E: ComplexField,

Computes the reconstructed matrix, given its QR decomposition, and stores the result in dst.

Panics

Panics if the number of columns of householder_factor isn’t the same as the minimum of the number of rows and the number of columns of qr_factors.

Panics if the block size is zero.

Panics if dst doesn’t have the same shape as qr_factors.

Panics if the provided memory in stack is insufficient (see reconstruct_req).