extendr_api::prelude::modules
Module svd

Low level implementation of the SVD of a matrix.

The SVD of a matrix $M$ of shape $(m, n)$ is a decomposition into three components $U$, $S$, and $V$, such that:

$U$ has shape $(m, m)$ and is a unitary matrix,

$V$ has shape $(n, n)$ and is a unitary matrix,

$S$ has shape $(m, n)$ and is zero everywhere except the main diagonal,

and finally:

$$M = U S V^H.$$

Structs

SvdParams

SVD tuning parameters.

Enums

ComputeVectors

Indicates whether the singular vectors are fully computed, partially computed, or skipped.

Functions

compute_svd

Computes the singular value decomposition of matrix.

compute_svd_custom_epsilon

See compute_svd.

compute_svd_req

Computes the size and alignment of required workspace for performing a singular value decomposition. $U$ and $V$ may be computed fully, partially, or not computed at all.