Type Definition extendr_api::prelude::prelude::CowArray

pub type CowArray<'a, A, D> = ArrayBase<CowRepr<'a, A>, D>;

An array with copy-on-write behavior.

An CowArray represents either a uniquely owned array or a view of an array. The 'a corresponds to the lifetime of the view variant.

This type is analogous to std::borrow::Cow. If a CowArray instance is the immutable view variant, then calling a method for mutating elements in the array will cause it to be converted into the owned variant (by cloning all the elements) before the modification is performed.

Array views have all the methods of an array (see ArrayBase).

See also ArcArray, which also provides copy-on-write behavior but has a reference-counted pointer to the data instead of either a view or a uniquely owned copy.