Type Definition extendr_api::prelude::prelude::RcArray

type RcArray<A, D> = ArrayBase<OwnedArcRepr<A>, D>;

👎 Deprecated:
RcArray has been renamed to ArcArray

An array where the data has shared ownership and is copy on write.

It can act as both an owner as the data as well as a shared reference (view like).

Note: this type alias is obsolete. See the equivalent ArcArray instead.