Trait extendr_api::prelude::row::RowIndex

pub trait RowIndex<ColRange>: Sized + Seal {
    type Target;

    // Required method
    fn get(this: Self, col: ColRange) -> Self::Target;

    // Provided method
    unsafe fn get_unchecked(this: Self, col: ColRange) -> Self::Target { ... }
}

Represents a type that can be used to slice a row, such as an index or a range of indices.

Required Associated Types

type Target

Resulting type of the indexing operation.

Required Methods

fn get(this: Self, col: ColRange) -> Self::Target

Index the row at col.

Provided Methods

unsafe fn get_unchecked(this: Self, col: ColRange) -> Self::Target

Index the row at col, without bound checks.

Object Safety

This trait is not object safe.