Struct extendr_api::prelude::solvers::Cholesky

pub struct Cholesky<E>
where
    E: Entity,
{
    factors: Mat<E>,
}

Cholesky decomposition.

Fields
factors: Mat<E>
Implementations

impl<E> Cholesky<E>
where E: ComplexField,

pub fn try_new<ViewE>( matrix: MatRef<'_, ViewE>, side: Side ) -> Result<Cholesky<E>, CholeskyError>
where ViewE: Conjugate<Canonical = E>,

Returns the Cholesky factorization of the input matrix, or an error if the matrix is not positive definite.

The factorization is such that $A = LL^H$, where $L$ is lower triangular.

The matrix is interpreted as Hermitian, but only the provided side is accessed.

pub fn compute_l(&self) -> Mat<E>

Returns the factor $L$ of the Cholesky decomposition.