extendr_api::prelude::solvers
Struct PartialPivLu

pub struct PartialPivLu<E>
where
    E: Entity,
{
    pub(crate) factors: Mat<E>,
    row_perm: Vec<usize>,
    row_perm_inv: Vec<usize>,
    n_transpositions: usize,
}

LU decomposition with partial pivoting.

Fields
factors: Mat<E>row_perm: Vec<usize>row_perm_inv: Vec<usize>n_transpositions: usize
Implementations

impl<E> PartialPivLu<E>
where E: ComplexField,

pub fn new<ViewE>(matrix: MatRef<'_, ViewE>) -> PartialPivLu<E>
where ViewE: Conjugate<Canonical = E>,

Returns the LU decomposition of the input matrix with partial (row) pivoting.

The factorization is such that $PA = LU$, where $L$ is lower triangular, $U$ is unit upper triangular, and $P$ is the permutation arising from the pivoting.

pub fn row_permutation(&self) -> PermRef<'_, usize>

Returns the row permutation due to pivoting.

pub fn transposition_count(&self) -> usize

Returns the number of transpositions that constitute the permutation.

pub fn compute_l(&self) -> Mat<E>

Returns the factor $L$ of the LU decomposition.

pub fn compute_u(&self) -> Mat<E>

Returns the factor $U$ of the LU decomposition.