Module extendr_api::prelude::sparse::linalg::cholesky::simplicial

Simplicial factorization module.

A simplicial factorization is one that processes the elements of the Cholesky factor of the input matrix single elements, rather than by blocks. This is more efficient if the Cholesky factor is very sparse.

Structs

EliminationTreeRef

Reference to a slice containing the Cholesky factor’s elimination tree.

SimplicialLdltRef

Cholesky LDLT factors containing both the symbolic and numeric representations.

SimplicialLltRef

Cholesky LLT factor containing both its symbolic and numeric representations.

SymbolicSimplicialCholesky

Cholesky factor structure containing its symbolic structure.

Functions

factorize_simplicial_numeric_ldlt

Computes the numeric values of the Cholesky LDLT factors of the matrix A, and stores them in L_values.

factorize_simplicial_numeric_ldlt_req

Returns the size and alignment of the workspace required to compute the numeric Cholesky LDLT factorization of a matrix A with dimension n.

factorize_simplicial_numeric_ldlt_with_row_indices

Computes the row indices and numeric values of the Cholesky LDLT factor of the matrix A, and stores them in L_row_indices and L_values.

factorize_simplicial_numeric_llt

Computes the numeric values of the Cholesky LLT factor of the matrix A, and stores them in L_values.

factorize_simplicial_numeric_llt_req

Returns the size and alignment of the workspace required to compute the numeric Cholesky LLT factorization of a matrix A with dimension n.

factorize_simplicial_numeric_llt_with_row_indices

Computes the row indices and numeric values of the Cholesky LLT factor of the matrix A, and stores them in L_row_indices and L_values.

factorize_simplicial_symbolic

Computes the symbolic structure of the Cholesky factor of the matrix A.

factorize_simplicial_symbolic_req

Computes the size and alignment of the workspace required to compute the symbolic Cholesky factorization of a square matrix with size n.

prefactorize_symbolic_cholesky

Computes the elimination tree and column counts of the Cholesky factorization of the matrix A.

prefactorize_symbolic_cholesky_req

Computes the size and alignment of the workspace required to compute the elimination tree and column counts of a matrix of size n with nnz non-zero entries.