extendr_api::prelude::sparse::linalg::cholesky::simplicial
Function factorize_simplicial_symbolic

pub fn factorize_simplicial_symbolic<I>(
    A: SymbolicSparseColMatRef<'_, I>,
    etree: EliminationTreeRef<'_, I>,
    col_counts: &[I],
    stack: PodStack<'_>,
) -> Result<SymbolicSimplicialCholesky<I>, FaerError>
where
    I: Index,

Computes the symbolic structure of the Cholesky factor of the matrix A.

Note

Only the upper triangular part of A is analyzed.

Panics

The elimination tree and column counts must be computed by calling prefactorize_symbolic_cholesky with the same matrix. Otherwise, the behavior is unspecified and panics may occur.