Struct extendr_api::prelude::sparse::linalg::cholesky::LltRef

pub struct LltRef<'a, I, E>
where
    I: Index,
    E: Entity,
{
    symbolic: &'a SymbolicCholesky<I>,
    values: SliceGroup<'a, E>,
}

Sparse LLT factorization wrapper.

Fields
symbolic: &'a SymbolicCholesky<I>values: SliceGroup<'a, E>
Implementations

impl<'a, I, E> LltRef<'a, I, E>
where I: Index, E: Entity,

pub fn new( symbolic: &'a SymbolicCholesky<I>, values: <<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>, ) -> LltRef<'a, I, E>

Creates a new Cholesky LLT factor from the symbolic part and numerical values.

Panics

Panics if values.len() != symbolic.len_values().

pub fn symbolic(self) -> &'a SymbolicCholesky<I>

Returns the symbolic part of the Cholesky factor.

pub fn solve_in_place_with_conj( &self, conj: Conj, rhs: MatMut<'_, E>, parallelism: Parallelism<'_>, stack: PodStack<'_>, )
where E: ComplexField,

Solves the equation $\text{Op}(A) x = \text{rhs}$ and stores the result in rhs, where $\text{Op}$ is either the identity or the conjugate, depending on the value of conj.

Panics

Panics if rhs.nrows() != self.symbolic().nrows().