extendr_api::prelude::sparse::linalg::cholesky::supernodal
Function factorize_supernodal_numeric_intranode_bunch_kaufman

pub fn factorize_supernodal_numeric_intranode_bunch_kaufman<I, E>(
    L_values: <<E as Entity>::Group as ForType>::FaerOf<&mut [<E as Entity>::Unit]>,
    subdiag: <<E as Entity>::Group as ForType>::FaerOf<&mut [<E as Entity>::Unit]>,
    perm_forward: &mut [I],
    perm_inverse: &mut [I],
    A_lower: SparseColMatRef<'_, I, E>,
    regularization: BunchKaufmanRegularization<'_, E>,
    symbolic: &SymbolicSupernodalCholesky<I>,
    parallelism: Parallelism<'_>,
    stack: PodStack<'_>,
) -> usize
where
    I: Index,
    E: ComplexField,

Computes the numeric values of the Cholesky Bunch-Kaufman factors of the matrix A with intranodal pivoting, and stores them in L_values.

Note

Only the lower (not upper, unlikely the other functions) triangular part of A is accessed.

Panics

The symbolic structure must be computed by calling factorize_supernodal_symbolic on a matrix with the same symbolic structure. Otherwise, the behavior is unspecified and panics may occur.