Function extendr_api::prelude::sparse::linalg::colamd::order

pub fn order<I>(
    perm: &mut [I],
    perm_inv: &mut [I],
    A: SymbolicSparseColMatRef<'_, I>,
    control: Control,
    stack: PodStack<'_>
) -> Result<(), FaerError>
where
    I: Index,

Computes the approximate minimum degree ordering for reducing the fill-in during the sparse QR factorization of a matrix with the sparsity pattern of A.

Note

Allows unsorted matrices.