extendr_api::prelude::sparse::linalg
Module lu

Computes the LU decomposition of a given sparse matrix. See faer::linalg::lu for more info.

The entry point in this module is SymbolicLu and factorize_symbolic_lu.

Warning

The functions in this module accept unsorted input, and always produce unsorted decomposition factors.

Modules

simplicial

Simplicial factorization module.

supernodal

Supernodal factorization module.

Structs

LuRef

Sparse LU factorization wrapper.

LuSymbolicParams

Tuning parameters for the LU symbolic factorization.

NumericLu

Structure that contains the numerical values and row pivoting permutation of the LU decomposition.

SymbolicLu

The symbolic structure of a sparse LU decomposition.

Enums

SymbolicLuRaw

The inner factorization used for the symbolic LU, either simplicial or symbolic.

Functions

factorize_symbolic_lu

Computes the symbolic LU factorization of the matrix A, or returns an error if the operation could not be completed.