Struct extendr_api::prelude::sparse::linalg::lu::simplicial::SimplicialLu

pub struct SimplicialLu<I, E>
where
    E: Entity,
{
    nrows: usize,
    ncols: usize,
    l_col_ptr: Vec<I>,
    l_row_ind: Vec<I>,
    l_val: VecGroup<E>,
    u_col_ptr: Vec<I>,
    u_row_ind: Vec<I>,
    u_val: VecGroup<E>,
}

LU factor structure containing the symbolic and numerical representations.

Fields
nrows: usizencols: usizel_col_ptr: Vec<I>l_row_ind: Vec<I>l_val: VecGroup<E>u_col_ptr: Vec<I>u_row_ind: Vec<I>u_val: VecGroup<E>
Implementations

impl<I, E> SimplicialLu<I, E>
where I: Index, E: Entity,

pub fn new() -> SimplicialLu<I, E>

Creates a new Simplicial LU of a $0\times 0$ matrix.

pub fn nrows(&self) -> usize

Returns the number of rows of $A$.

pub fn ncols(&self) -> usize

Returns the number of columns of $A$.

pub fn l_factor_unsorted(&self) -> SparseColMatRef<'_, I, E>

Returns the $L$ factor of the LU factorization. The row indices may or may not be sorted.

pub fn u_factor_unsorted(&self) -> SparseColMatRef<'_, I, E>

Returns the $U$ factor of the LU factorization. The row indices may or may not be sorted.

pub fn solve_in_place_with_conj( &self, row_perm: PermRef<'_, I>, col_perm: PermRef<'_, I>, conj_lhs: Conj, rhs: MatMut<'_, E>, parallelism: Parallelism, work: MatMut<'_, E>, )
where E: ComplexField,

Solves the equation $\text{Op}(A) x = \text{rhs}$ and stores the result in rhs, where $\text{Op}$ is either the identity or the conjugate, depending on the value of conj.

Panics

Panics if self.nrows() != self.ncols().

Panics if rhs.nrows() != self.nrows().

pub fn solve_transpose_in_place_with_conj( &self, row_perm: PermRef<'_, I>, col_perm: PermRef<'_, I>, conj_lhs: Conj, rhs: MatMut<'_, E>, parallelism: Parallelism, work: MatMut<'_, E>, )
where E: ComplexField,

Solves the equation $\text{Op}(A^\top) x = \text{rhs}$ and stores the result in rhs, where $\text{Op}$ is either the identity or the conjugate, depending on the value of conj.

Panics

Panics if self.nrows() != self.ncols().

Panics if rhs.nrows() != self.nrows().