Module extendr_api::prelude::sparse::linalg::matmul

Sparse matrix multiplication.

Structs

SparseMatmulInfo

Info about the matrix multiplication operation to help split the workload between multiple threads.

Functions

dense_sparse_matmul

Multiplies a dense matrix lhs by a sparse matrix rhs, and stores the result in acc. See faer::linalg::matmul::matmul for more details.

sparse_dense_matmul

Multiplies a sparse matrix lhs by a dense matrix rhs, and stores the result in acc. See faer::linalg::matmul::matmul for more details.

sparse_sparse_matmul

Multiplies a sparse matrix lhs by a sparse matrix rhs, multiplied by k, and returns the result.

sparse_sparse_matmul_numeric

Performs a numeric matrix multiplication of a sparse matrix lhs by a sparse matrix rhs multiplied by k, and stores the result in dst.

sparse_sparse_matmul_numeric_req

Computes the size and alignment of the workspace required to perform the numeric matrix multiplication into dst.

sparse_sparse_matmul_symbolic

Performs a symbolic matrix multiplication of a sparse matrix lhs by a sparse matrix rhs, and returns the result.