extendr_api::prelude::sparse::linalg::matmul
Function sparse_sparse_matmul_symbolic

pub fn sparse_sparse_matmul_symbolic<I>(
    lhs: SymbolicSparseColMatRef<'_, I>,
    rhs: SymbolicSparseColMatRef<'_, I>,
) -> Result<(SymbolicSparseColMat<I>, SparseMatmulInfo), FaerError>
where
    I: Index,

Performs a symbolic matrix multiplication of a sparse matrix lhs by a sparse matrix rhs, and returns the result.

Note

Allows unsorted matrices, and produces a sorted output.