extendr_api::prelude::sparse::linalg::qr
Module simplicial

Simplicial factorization module.

A simplicial factorization is one that processes the elements of the QR factors of the input matrix by single elements, rather than by blocks. This is more efficient if the QR factors are very sparse.

Structs

SimplicialQrRef

QR factors containing both the symbolic and numeric representations.

SymbolicSimplicialQr

Symbolic structure of the QR decomposition,

Functions

factorize_simplicial_numeric_qr_req

Computes the size and alignment of the workspace required to compute the numerical QR factorization of the matrix whose structure was used to produce the symbolic structure.

factorize_simplicial_numeric_qr_unsorted

Computes the numerical QR factorization of $A$.

factorize_simplicial_symbolic_qr

Computes the symbolic QR factorization of a matrix $A$, given the outputs of the pre-factorization steps.

factorize_simplicial_symbolic_qr_req

Computes the size and alignment of the workspace required to compute the symbolic QR factorization of a matrix with dimensions (nrows, ncols).