Module extendr_api::prelude::sparse::linalg::solvers

High level sparse solvers.

Structs

Cholesky

Sparse Cholesky factorization.

Lu

Sparse LU factorization.

Qr

Sparse QR factorization.

SymbolicCholesky

Reference-counted sparse symbolic Cholesky factorization.

SymbolicLu

Reference-counted sparse symbolic LU factorization.

SymbolicQr

Reference-counted sparse symbolic QR factorization.

Traits

SpSolver

Solver that can compute solution of a linear system.

SpSolverCore

Object-safe base for SpSolver

SpSolverLstsq

Solver that can compute the least squares solution of an overdetermined linear system.

SpSolverLstsqCore

Object-safe base for SpSolverLstsq