Module extendr_api::prelude::sparse::linalg::triangular_solve

Solving sparse triangular linear systems with a dense right-hand-side.

Functions

solve_lower_triangular_in_place

Assuming self is a lower triangular matrix, solves the equation Op(self) * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

solve_lower_triangular_transpose_in_place

Assuming self is a lower triangular matrix, solves the equation Op(self).transpose() * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

solve_unit_lower_triangular_in_place

Assuming self is a unit lower triangular matrix, solves the equation Op(self) * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

solve_unit_lower_triangular_transpose_in_place

Assuming self is a unit lower triangular matrix, solves the equation Op(self).transpose() * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

solve_unit_upper_triangular_in_place

Assuming self is a unit upper triangular matrix, solves the equation Op(self) * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

solve_unit_upper_triangular_transpose_in_place

Assuming self is a unit upper triangular matrix, solves the equation Op(self).transpose() * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

solve_upper_triangular_in_place

Assuming self is an upper triangular matrix, solves the equation Op(self) * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

solve_upper_triangular_transpose_in_place

Assuming self is an upper triangular matrix, solves the equation Op(self).transpose() * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.