extendr_api::prelude::sparse::linalg::triangular_solve
Function solve_unit_lower_triangular_transpose_in_place

pub fn solve_unit_lower_triangular_transpose_in_place<I, E>(
    l: SparseColMatRef<'_, I, E>,
    conj: Conj,
    rhs: MatMut<'_, E>,
    parallelism: Parallelism<'_>,
)
where
    I: Index,
    E: ComplexField,

Assuming self is a unit lower triangular matrix, solves the equation Op(self).transpose() * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

Note

The matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each column.