Function extendr_api::prelude::sparse::linalg::triangular_solve::solve_upper_triangular_in_place

pub fn solve_upper_triangular_in_place<I, E>(
    u: SparseColMatRef<'_, I, E>,
    conj: Conj,
    rhs: MatMut<'_, E>,
    parallelism: Parallelism
)
where
    I: Index,
    E: ComplexField,

Assuming self is an upper triangular matrix, solves the equation Op(self) * X = rhs, and stores the result in rhs, where Op is either the conjugate or the identity depending on the value of conj.

Note

The matrix indices need not be sorted, but the diagonal element is assumed to be the last stored element in each column.