Function extendr_api::prelude::sparse::ops::add_assign

pub fn add_assign<I, E, RhsE>(
    dst: SparseColMatMut<'_, I, E>,
    rhs: SparseColMatRef<'_, I, RhsE>,
)
where
    I: Index,
    E: ComplexField,
    RhsE: Conjugate<Canonical = E>,

Computes the sum of dst and src and stores the result in dst without changing its symbolic structure.

Panics

Panics if dst and rhs don’t have matching dimensions.
Panics if rhs contains an index that’s unavailable in dst.