extendr_api::prelude::sparse::ops
Function binary_op_assign_into

pub fn binary_op_assign_into<I, E, SrcE>(
    dst: SparseColMatMut<'_, I, E>,
    src: SparseColMatRef<'_, I, SrcE>,
    f: impl FnMut(E, SrcE) -> E,
)
where
    I: Index,
    E: Entity,
    SrcE: Entity,

Returns the resulting matrix obtained by applying f to the elements from dst and src skipping entries that are unavailable in both of them.
The sparsity patter of dst is unchanged.

Panics

Panics if src and dst don’t have matching dimensions.
Panics if src contains an index that’s unavailable in dst.