Module extendr_api::prelude::sparse::solvers

👎Deprecated: moved to faer::linalg::solvers

Sparse solvers.

Each solver satisfies the SpSolver and/or SpSolverLstsq traits, which can be used to solve linear systems.

Structs

Cholesky

Cholesky decomposition.

CholeskyError

This error signifies that the LLT decomposition could not be computed due to the matrix not being numerically positive definite.

ColPivQr

QR decomposition with column pivoting.

Eigendecomposition

Complex eigendecomposition.

FullPivLu

LU decomposition with full pivoting.

Lblt

Bunch-Kaufman decomposition.

PartialPivLu

LU decomposition with partial pivoting.

Qr

QR decomposition.

SelfAdjointEigendecomposition

Self-adjoint eigendecomposition.

Svd

Singular value decomposition.

ThinSvd

Thin singular value decomposition.

Traits

Solver

Solver that can compute solution of a linear system.

SolverCore

Object-safe base for Solver

SolverLstsq

Dense solver that can compute the least squares solution of an overdetermined linear system.

SolverLstsqCore

Object-safe base for SolverLstsq

SpSolver

Solver that can compute solution of a linear system.

SpSolverCore

Object-safe base for SpSolver

SpSolverLstsq

Solver that can compute the least squares solution of an overdetermined linear system.

SpSolverLstsqCore

Object-safe base for SpSolverLstsq