Trait extendr_api::prelude::sparse::solvers::SpSolverCore

pub trait SpSolverCore<E>
where
    E: Entity,
{
    // Required methods
    fn nrows(&self) -> usize;
    fn ncols(&self) -> usize;
}

Object-safe base for SpSolver

Required Methods

fn nrows(&self) -> usize

Returns the number of rows of the matrix used to construct this decomposition.

fn ncols(&self) -> usize

Returns the number of columns of the matrix used to construct this decomposition.