Struct extendr_api::prelude::sparse::SparseRowMatRef

pub struct SparseRowMatRef<'a, I, E>
where
    I: Index,
    E: Entity,
{
    pub(crate) symbolic: SymbolicSparseRowMatRef<'a, I>,
    pub(crate) values: SliceGroup<'a, E>,
}

Sparse matrix view in column-major format, either compressed or uncompressed.

Fields
symbolic: SymbolicSparseRowMatRef<'a, I>values: SliceGroup<'a, E>
Implementations

impl<'a, I, E> SparseRowMatRef<'a, I, E>
where I: Index, E: Entity,

pub fn new( symbolic: SymbolicSparseRowMatRef<'a, I>, values: <<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>, ) -> SparseRowMatRef<'a, I, E>

Creates a new sparse matrix view.

Panics

Panics if the length of values is not equal to the length of symbolic.col_indices().

pub fn nrows(&self) -> usize

Returns the number of rows of the matrix.

pub fn ncols(&self) -> usize

Returns the number of columns of the matrix.

pub fn as_ref(&self) -> SparseRowMatRef<'_, I, E>

Returns a view over self.

pub fn values( self, ) -> <<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>

Returns the numerical values of the matrix.

pub fn to_owned( &self, ) -> Result<SparseRowMat<I, <E as Conjugate>::Canonical>, FaerError>
where E: Conjugate, <E as Conjugate>::Canonical: ComplexField,

Copies self into a newly allocated matrix.

Note

Allows unsorted matrices, producing an unsorted output.

pub fn to_sorted( &self, ) -> Result<SparseRowMat<I, <E as Conjugate>::Canonical>, FaerError>
where E: Conjugate, <E as Conjugate>::Canonical: ComplexField,

Copies self into a newly allocated matrix with sorted indices.

Note

Allows unsorted matrices, producing a sorted output.

pub fn to_dense(&self) -> Mat<<E as Conjugate>::Canonical>
where E: Conjugate, <E as Conjugate>::Canonical: ComplexField,

Copies self into a newly allocated dense matrix

pub fn to_col_major( &self, ) -> Result<SparseColMat<I, <E as Conjugate>::Canonical>, FaerError>
where E: Conjugate, <E as Conjugate>::Canonical: ComplexField,

Copies self into a newly allocated matrix, with column-major order.

Note

Allows unsorted matrices, producing a sorted output.

pub fn transpose(self) -> SparseColMatRef<'a, I, E>

Returns a view over the transpose of self in column-major format.

pub fn conjugate(self) -> SparseRowMatRef<'a, I, <E as Conjugate>::Conj>
where E: Conjugate,

Returns a view over the conjugate of self.

pub fn canonicalize( self, ) -> (SparseRowMatRef<'a, I, <E as Conjugate>::Canonical>, Conj)
where E: Conjugate,

Returns a view over the conjugate of self.

pub fn adjoint(self) -> SparseColMatRef<'a, I, <E as Conjugate>::Conj>
where E: Conjugate,

Returns a view over the conjugate transpose of self.

pub fn values_of_row( self, i: usize, ) -> <<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>

Returns the numerical values of row i of the matrix.

Panics:

Panics if i >= nrows.

pub fn symbolic(&self) -> SymbolicSparseRowMatRef<'a, I>

Returns the symbolic structure of the matrix.

pub fn into_parts( self, ) -> (SymbolicSparseRowMatRef<'a, I>, <<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>)

Decomposes the matrix into the symbolic part and the numerical values.

pub fn compute_nnz(&self) -> usize

Returns the number of symbolic non-zeros in the matrix.

The value is guaranteed to be less than I::Signed::MAX.

Note

Allows unsorted matrices, but the output is a count of all the entries, including the duplicate ones.

pub fn row_ptrs(&self) -> &'a [I]

Returns the column pointers.

pub fn nnz_per_row(&self) -> Option<&'a [I]>

Returns the count of non-zeros per column of the matrix.

pub fn col_indices(&self) -> &'a [I]

Returns the column indices.

pub fn col_indices_of_row_raw(&self, i: usize) -> &'a [I]

Returns the column indices of row i.

Panics

Panics if i >= self.nrows().

pub fn col_indices_of_row( &self, i: usize, ) -> impl ExactSizeIterator + DoubleEndedIterator + 'a

Returns the column indices of row i.

Panics

Panics if i >= self.ncols().

pub fn row_range(&self, i: usize) -> Range<usize>

Returns the range that the row i occupies in self.col_indices().

Panics

Panics if i >= self.nrows().

pub unsafe fn row_range_unchecked(&self, i: usize) -> Range<usize>

Returns the range that the row i occupies in self.col_indices().

Safety

The behavior is undefined if i >= self.nrows().

pub fn get( self, row: usize, col: usize, ) -> Option<<<E as Entity>::Group as ForType>::FaerOf<&'a <E as Entity>::Unit>>

Returns a reference to the value at the given index, or None if the symbolic structure doesn’t contain it, or contains multiple values with the given index.

Panics

Panics if row >= self.nrows()
Panics if col >= self.ncols()

pub fn get_all( self, row: usize, col: usize, ) -> <<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>

Returns a reference to a slice containing the values at the given index using a binary search.

Panics

Panics if row >= self.nrows().
Panics if col >= self.ncols().

impl<I, E> SparseRowMatRef<'_, I, E>
where I: Index, E: ComplexField,

pub fn sp_solve_lower_triangular_in_place(&self, rhs: impl ColBatchMut<E>)

Assuming self is an upper triangular matrix, solves the equation self * X = rhs, and stores the result in rhs.

Note

The matrix indices need not be sorted, but the diagonal element is assumed to be the last stored element in each row.

pub fn sp_solve_upper_triangular_in_place(&self, rhs: impl ColBatchMut<E>)

Assuming self is an upper triangular matrix, solves the equation self * X = rhs, and stores the result in rhs.

Note

The matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each row.

pub fn sp_solve_unit_lower_triangular_in_place(&self, rhs: impl ColBatchMut<E>)

Assuming self is a unit lower triangular matrix, solves the equation self * X = rhs, and stores the result in rhs.

Note

The matrix indices need not be sorted, but the diagonal element is assumed to be the last stored element in each row.

pub fn sp_solve_unit_upper_triangular_in_place(&self, rhs: impl ColBatchMut<E>)

Assuming self is a unit upper triangular matrix, solves the equation self * X = rhs, and stores the result in rhs.

Note

The matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each row.

pub fn sp_cholesky(&self, side: Side) -> Result<Cholesky<I, E>, CholeskyError>

Returns the Cholesky decomposition of self. Only the provided side is accessed.

pub fn sp_lu(&self) -> Result<Lu<I, E>, LuError>

Returns the LU decomposition of self with partial (row) pivoting.

pub fn sp_qr(&self) -> Result<Qr<I, E>, FaerError>

Returns the QR decomposition of self.