Module extendr_api::prelude::sparse::utils

Useful sparse matrix primitives.

Functions

adjoint

Computes the adjoint of the matrix A and returns a view over it.

adjoint_symbolic

Computes the adjoint of the symbolic matrix A and returns a view over it.

permute_hermitian

Computes the self-adjoint permutation $P A P^\top$ of the matrix A and returns a view over it.

permute_hermitian_req

Computes the size and alignment of the workspace required to compute a two-sided permutation of a self-adjoint matrix.

sort_dedup_indices

Sorts and deduplicates row_indices and values simultaneously so that row_indices is nonincreasing and contains no duplicate indices.

sort_indices

Sorts row_indices and values simultaneously so that row_indices is nonincreasing.

transpose

Computes the transpose of the matrix A and returns a view over it.