Struct extendr_api::prelude::RowMut

#[repr(C)]
pub struct RowMut<'a, E>
where
    E: Entity,
{
    pub(super) inner: VecImpl<E>,
    pub(super) __marker: PhantomData<&'a E>,
}

Mutable view over a row vector, similar to a mutable reference to a strided slice.

Note

Unlike a slice, the data pointed to by RowMut<'_, E> is allowed to be partially or fully uninitialized under certain conditions. In this case, care must be taken to not perform any operations that read the uninitialized values, or form references to them, either directly through RowMut::read, or indirectly through any of the numerical library routines, unless it is explicitly permitted.

Move semantics

See faer::Mat for information about reborrowing when using this type.

Fields
inner: VecImpl<E>__marker: PhantomData<&'a E>
Implementations

impl<'a, E> RowMut<'a, E>
where E: Entity,

pub fn nrows(&self) -> usize

Returns the number of rows of the row. This is always equal to 1.

pub fn ncols(&self) -> usize

Returns the number of columns of the row.

pub fn as_ptr_mut( self, ) -> <<E as Entity>::Group as ForType>::FaerOf<*mut <E as Entity>::Unit>

Returns pointers to the matrix data.

pub fn col_stride(&self) -> isize

Returns the column stride of the matrix, specified in number of elements, not in bytes.

pub fn as_2d_mut(self) -> MatMut<'a, E>

Returns self as a mutable matrix view.

pub fn ptr_at_mut( self, col: usize, ) -> <<E as Entity>::Group as ForType>::FaerOf<*mut <E as Entity>::Unit>

Returns raw pointers to the element at the given index.

pub unsafe fn ptr_inbounds_at_mut( self, col: usize, ) -> <<E as Entity>::Group as ForType>::FaerOf<*mut <E as Entity>::Unit>

Returns raw pointers to the element at the given index, assuming the provided index is within the size of the vector.

Safety

The behavior is undefined if any of the following conditions are violated:

col < self.ncols().

pub unsafe fn split_at_mut_unchecked( self, col: usize, ) -> (RowMut<'a, E>, RowMut<'a, E>)

Splits the column vector at the given index into two parts and returns an array of each subvector, in the following order:

left.

right.

Safety

The behavior is undefined if any of the following conditions are violated:

col <= self.ncols().

pub fn split_at_mut(self, col: usize) -> (RowMut<'a, E>, RowMut<'a, E>)

Splits the column vector at the given index into two parts and returns an array of each subvector, in the following order:

top.

bottom.

Panics

The function panics if any of the following conditions are violated:

col <= self.ncols().

pub unsafe fn get_mut_unchecked<ColRange>( self, col: ColRange, ) -> <RowMut<'a, E> as RowIndex<ColRange>>::Target
where RowMut<'a, E>: RowIndex<ColRange>,

Returns references to the element at the given index, or subvector if col is a range.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

Safety

The behavior is undefined if any of the following conditions are violated:

col must be contained in [0, self.ncols()).

pub fn get_mut<ColRange>( self, col: ColRange, ) -> <RowMut<'a, E> as RowIndex<ColRange>>::Target
where RowMut<'a, E>: RowIndex<ColRange>,

Returns references to the element at the given index, or subvector if col is a range, with bound checks.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

Panics

The function panics if any of the following conditions are violated:

col must be contained in [0, self.ncols()).

pub unsafe fn read_unchecked(&self, col: usize) -> E

Reads the value of the element at the given index.

Safety

The behavior is undefined if any of the following conditions are violated:

col < self.ncols().

pub fn read(&self, col: usize) -> E

Reads the value of the element at the given index, with bound checks.

Panics

The function panics if any of the following conditions are violated:

col < self.ncols().

pub unsafe fn write_unchecked(&mut self, col: usize, value: E)

Writes the value to the element at the given index.

Safety

The behavior is undefined if any of the following conditions are violated:

col < self.ncols().

pub fn write(&mut self, col: usize, value: E)

Writes the value to the element at the given index, with bound checks.

Panics

The function panics if any of the following conditions are violated:

col < self.ncols().

pub fn copy_from(&mut self, other: impl AsRowRef<E>)

Copies the values from other into self.

Panics

The function panics if any of the following conditions are violated:

self.ncols() == other.ncols().

pub fn fill_zero(&mut self)
where E: ComplexField,

Fills the elements of self with zeros.

pub fn fill(&mut self, constant: E)

Fills the elements of self with copies of constant.

pub fn transpose_mut(self) -> ColMut<'a, E>

Returns a view over the transpose of self.

pub fn conjugate_mut(self) -> RowMut<'a, <E as Conjugate>::Conj>
where E: Conjugate,

Returns a view over the conjugate of self.

pub fn adjoint_mut(self) -> ColMut<'a, <E as Conjugate>::Conj>
where E: Conjugate,

Returns a view over the conjugate transpose of self.

pub fn canonicalize_mut(self) -> (RowMut<'a, <E as Conjugate>::Canonical>, Conj)
where E: Conjugate,

Returns a view over the canonical representation of self, as well as a flag declaring whether self is implicitly conjugated or not.

pub fn reverse_cols_mut(self) -> RowMut<'a, E>

Returns a view over the self, with the columnss in reversed order.

pub unsafe fn subcols_mut_unchecked( self, col_start: usize, ncols: usize, ) -> RowMut<'a, E>

Returns a view over the subvector starting at col col_start, and with number of columns ncols.

Safety

The behavior is undefined if any of the following conditions are violated:

col_start <= self.ncols().

ncols <= self.ncols() - col_start.

pub fn subcols_mut(self, col_start: usize, ncols: usize) -> RowMut<'a, E>

Returns a view over the subvector starting at col col_start, and with number of columns ncols.

Safety

The behavior is undefined if any of the following conditions are violated:

col_start <= self.ncols().

ncols <= self.ncols() - col_start.

pub fn to_owned(&self) -> Row<<E as Conjugate>::Canonical>
where E: Conjugate,

Returns an owning Row of the data.

pub fn has_nan(&self) -> bool
where E: ComplexField,

Returns true if any of the elements is NaN, otherwise returns false.

pub fn is_all_finite(&self) -> bool
where E: ComplexField,

Returns true if all of the elements are finite, otherwise returns false.

pub fn norm_max(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the maximum norm of self.

pub fn norm_l1(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the L1 norm of self.

pub fn norm_l2(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the L2 norm of self.

pub fn squared_norm_l2(&self) -> <E as ComplexField>::Real
where E: ComplexField,

Returns the squared L2 norm of self.

pub fn sum(&self) -> E
where E: ComplexField,

Returns the sum of self.

pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
where E: ComplexField,

Kroneckor product of self and rhs.

This is an allocating operation; see faer::linalg::kron for the allocation-free version or more info in general.

pub fn try_as_slice( self, ) -> Option<<<E as Entity>::Group as ForType>::FaerOf<&'a [<E as Entity>::Unit]>>

Returns the row as a contiguous slice if its column stride is equal to 1.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

pub fn try_as_slice_mut( self, ) -> Option<<<E as Entity>::Group as ForType>::FaerOf<&'a mut [<E as Entity>::Unit]>>

Returns the row as a contiguous slice if its column stride is equal to 1.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

pub unsafe fn try_as_uninit_slice_mut( self, ) -> Option<<<E as Entity>::Group as ForType>::FaerOf<&'a mut [MaybeUninit<<E as Entity>::Unit>]>>

Returns the row as a contiguous potentially uninitialized slice if its column stride is equal to 1.

Safety

If uninit data is written to the slice, it must not be later read.

pub fn as_ref(&self) -> RowRef<'_, E>

Returns a view over the matrix.

pub fn as_mut(&mut self) -> RowMut<'_, E>

Returns a mutable view over the matrix.