Trait extendr_api::prelude::AssignElem

pub trait AssignElem<T> {
    fn assign_elem(self, input: T);
}

A producer element that can be assigned to once

Required methods

fn assign_elem(self, input: T)

Assign the value input to the element that self represents.