Trait extendr_api::prelude::MultiSliceArg

pub trait MultiSliceArg<'a, A, D> where
    A: 'a,
    D: Dimension, {
    type Output;
    fn multi_slice_move(
        &self, 
        view: ArrayBase<ViewRepr<&'a mut A>, D>
    ) -> Self::Output;
    fn __private__(&self) -> PrivateMarker;
}

Slicing information describing multiple mutable, disjoint slices.

It’s unfortunate that we need 'a and A to be parameters of the trait, but they’re necessary until Rust supports generic associated types.

Associated Types

type Output

The type of the slices created by .multi_slice_move().

Required methods

fn multi_slice_move(
&self,
view: ArrayBase<ViewRepr<&'a mut A>, D>
) -> Self::Output

Split the view into multiple disjoint slices.

Panics if performing any individual slice panics or if the slices are not disjoint (i.e. if they intersect).

fn __private__(&self) -> PrivateMarker

This trait is private to implement; this method exists to make it impossible to implement outside the crate.