Trait extendr_api::prelude::RawDataSubst

pub trait RawDataSubst<A>: RawData {
    type Output: RawData<Elem = A>;

    unsafe fn data_subst(self) -> Self::Output;
}

Array representation trait.

The RawDataSubst trait maps the element type of array storage, while keeping the same kind of storage.

For example, RawDataSubst<B> can map the type OwnedRepr<A> to OwnedRepr<B>.

Required Associated Types

type Output: RawData<Elem = A>

The resulting array storage of the same kind but substituted element type

Required Methods

unsafe fn data_subst(self) -> Self::Output

Unsafely translate the data representation from one element representation to another.

Safety

Caller must ensure the two types have the same representation.