extendr_api::prelude::utils::constrained::mat
Struct MatMut

#[repr(transparent)]
pub struct MatMut<'nrows, 'ncols, 'a, E>(Branded<'ncols, Branded<'nrows, MatMut<'a, E>>>)
where
    E: Entity;

Mutable dense matrix view with dimensions equal to the values tied to ('nrows, 'ncols).

Tuple Fields
0: Branded<'ncols, Branded<'nrows, MatMut<'a, E>>>
Implementations

impl<'nrows, 'ncols, 'a, E> MatMut<'nrows, 'ncols, 'a, E>
where E: Entity,

pub fn new( inner: MatMut<'a, E>, nrows: Size<'nrows>, ncols: Size<'ncols>, ) -> MatMut<'nrows, 'ncols, 'a, E>

Returns a new matrix view after checking that its dimensions match the dimensions tied to ('nrows, 'ncols).

pub fn nrows(&self) -> Size<'nrows>

Returns the number of rows of the matrix.

pub fn ncols(&self) -> Size<'ncols>

Returns the number of columns of the matrix.

pub fn into_inner(self) -> MatMut<'a, E>

Returns the unconstrained matrix.

pub fn read(&self, i: Idx<'nrows, usize>, j: Idx<'ncols, usize>) -> E

Returns the element at position (i, j).

pub fn write(&mut self, i: Idx<'nrows, usize>, j: Idx<'ncols, usize>, value: E)

Writes value to the location at position (i, j).