Struct extendr_api::prelude::utils::constrained::sparse::SparseColMatMut

pub struct SparseColMatMut<'nrows, 'ncols, 'a, I, E>
where
    I: Index,
    E: Entity,
{
    symbolic: SymbolicSparseColMatRef<'nrows, 'ncols, 'a, I>,
    values: SliceGroupMut<'a, E>,
}

Mutable sparse matrix view with dimensions equal to the values tied to ('nrows, 'ncols), in column-major order.

Fields
symbolic: SymbolicSparseColMatRef<'nrows, 'ncols, 'a, I>values: SliceGroupMut<'a, E>
Implementations

impl<'nrows, 'ncols, 'a, I, E> SparseColMatMut<'nrows, 'ncols, 'a, I, E>
where I: Index, E: Entity,

pub fn new( inner: SparseColMatMut<'a, I, E>, nrows: Size<'nrows>, ncols: Size<'ncols>, ) -> SparseColMatMut<'nrows, 'ncols, 'a, I, E>

Returns a new matrix view after checking that its dimensions match the dimensions tied to ('nrows, 'ncols).

pub fn into_inner(self) -> SparseColMatMut<'a, I, E>

Returns the unconstrained matrix.

pub fn values_of_col_mut( &mut self, j: Idx<'ncols, usize>, ) -> <<E as Entity>::Group as ForType>::FaerOf<&mut [<E as Entity>::Unit]>

Returns the values in column j.

pub fn symbolic(&self) -> SymbolicSparseColMatRef<'nrows, 'ncols, 'a, I>

Returns the symbolic structure of the matrix.