Struct extendr_api::prelude::utils::constrained::Size

#[repr(transparent)]
pub struct Size<'n>(Branded<'n, usize>);

usize value tied to the lifetime 'n.

Tuple Fields
0: Branded<'n, usize>
Implementations

impl<'size> Size<'size>

pub fn with<R>(n: usize, f: impl for<'n> FnOnce(Size<'n>) -> R) -> R

Create a new Size with a lifetime tied to n.

pub fn with2<R>( m: usize, n: usize, f: impl for<'m, 'n> FnOnce(Size<'m>, Size<'n>) -> R, ) -> R

Create two new Size with lifetimes tied to m and n.

pub unsafe fn new_raw_unchecked(n: usize) -> Size<'size>

Create a new Size tied to the lifetime 'n.

pub fn into_inner(self) -> usize

Returns the unconstrained value.

pub fn indices(self) -> impl DoubleEndedIterator

Returns an iterator of the indices smaller than self.

pub fn check<I>(self, idx: I) -> Idx<'size, I>
where I: Index,

Check that the index is bounded by self, or panic otherwise.

pub fn try_check<I>(self, idx: I) -> Option<Idx<'size, I>>
where I: Index,

Check that the index is bounded by self, or return None otherwise.