Module extendr_api::prelude::utils::slice

Slice types for entities.

Structs

RefGroup

Wrapper around a group of references.

RefGroupMut

Wrapper around a group of mutable references.

SliceGroup

Analogous to an immutable reference to a slice for groups.

SliceGroupMut

Analogous to a mutable reference to a slice for groups.