extendr_api::prelude::utils::slice
Struct SliceGroupMut

pub struct SliceGroupMut<'a, E, T = <E as Entity>::Unit>(<<E as Entity>::Group as ForType>::FaerOf<*mut [T]>, PhantomData<&'a mut ()>)
where
    E: Entity,
    T: 'a;

Analogous to a mutable reference to a slice for groups.

Tuple Fields
0: <<E as Entity>::Group as ForType>::FaerOf<*mut [T]>1: PhantomData<&'a mut ()>
Implementations

impl<'a, E, T> SliceGroupMut<'a, E, T>
where E: Entity,

pub fn new( slice: <<E as Entity>::Group as ForType>::FaerOf<&'a mut [T]>, ) -> SliceGroupMut<'a, E, T>

Create a new SliceGroup from a group of mutable slice references.

pub fn into_inner( self, ) -> <<E as Entity>::Group as ForType>::FaerOf<&'a mut [T]>

Consume self to return the internally stored group of mutable slice references.

pub fn as_arrays_mut<const N: usize>( self, ) -> (SliceGroupMut<'a, E, [T; N]>, SliceGroupMut<'a, E, T>)

Decompose self into a mutable slice of arrays of size N, and a remainder part with length < N.

impl<'a, E> SliceGroupMut<'a, E>
where E: Entity,

pub fn read(&self, idx: usize) -> E

Read the element at position idx.

pub unsafe fn read_unchecked(&self, idx: usize) -> E

Read the element at position idx, without bound checks.

Safety

The behavior is undefined if idx >= self.len().

pub fn write(&mut self, idx: usize, value: E)

Write value to the location at position idx.

pub unsafe fn write_unchecked(&mut self, idx: usize, value: E)

Write value to the location at position idx, without bound checks.

Safety

The behavior is undefined if idx >= self.len().

pub fn fill_zero(&mut self)

Fill the slice with zeros.

impl<'a, E, T> SliceGroupMut<'a, E, T>
where E: Entity,

pub fn get_mut(self, idx: usize) -> RefGroupMut<'a, E, T>

Get a RefGroupMut pointing to the element at position idx.

pub unsafe fn get_unchecked_mut(self, idx: usize) -> RefGroupMut<'a, E, T>

Get a RefGroupMut pointing to the element at position idx.

Safety

The behavior is undefined if idx >= self.len().

pub fn get(self, idx: usize) -> RefGroup<'a, E, T>

Get a RefGroup pointing to the element at position idx.

pub unsafe fn get_unchecked(self, idx: usize) -> RefGroup<'a, E, T>

Get a RefGroup pointing to the element at position idx, without bound checks.

Safety

The behavior is undefined if idx >= self.len().

pub fn is_empty(&self) -> bool

Checks whether the slice is empty.

pub fn len(&self) -> usize

Returns the length of the slice.

pub fn subslice(self, range: Range<usize>) -> SliceGroupMut<'a, E, T>

Returns the subslice of self from the start to the end of the provided range.

pub unsafe fn subslice_unchecked( self, range: Range<usize>, ) -> SliceGroupMut<'a, E, T>

Returns the subslice of self from the start to the end of the provided range, without bound checks.

Safety

The behavior is undefined if range.start > range.end or range.end > self.len().

pub fn into_mut_iter(self) -> impl Iterator<Item = RefGroupMut<'a, E, T>>

Returns an iterator of RefGroupMut over the elements of the slice.

pub fn split_at( self, idx: usize, ) -> (SliceGroupMut<'a, E, T>, SliceGroupMut<'a, E, T>)

Split self at the midpoint idx, and return the two parts.

pub fn into_chunks_exact( self, chunk_size: usize, ) -> (impl Iterator<Item = SliceGroupMut<'a, E, T>>, SliceGroupMut<'a, E, T>)

Returns an iterator of slices over chunks of size chunk_size, and the remainder of the slice.