Function extendr_api::prelude::utils::thread::for_each_raw

pub fn for_each_raw(
    n_tasks: usize,
    op: impl Send + Sync + Fn(usize),
    parallelism: Parallelism,
)

Executes the tasks by passing the values in 0..n_tasks to op, possibly in parallel, while splitting the amount of parallelism between the tasks.