extendr_api::robj

Module into_robj

Modules

repeat_into_robj 🔒

Macros

impl_complex_tvv 🔒

impl_from_as_iterator 🔒

impl_integer_tvv 🔒

impl_real_tvv 🔒

impl_str_tvv 🔒

Traits

IntoRobj

This is an extension trait to provide a convenience method into_robj().

RobjItertools

Extensions to iterators for R objects including RobjItertools::collect_robj().

ToVectorValue

ToVectorValue is a trait that allows many different types to be converted to vectors. It is used as a type parameter to collect_robj().

Functions

fixed_size_collect 🔒

str_to_character 🔒

Returns an CHARSXP based on the provided &str.