Trait extendr_api::robj::Slices

pub trait Slices: GetSexp {
    // Provided methods
    unsafe fn as_typed_slice_raw<T>(&self) -> &[T] { ... }
    unsafe fn as_typed_slice_raw_mut<T>(&mut self) -> &mut [T] { ... }
}

Provided Methods

unsafe fn as_typed_slice_raw<T>(&self) -> &[T]

Get an immutable slice to this object’s data.

Safety

Unless the type is correct, this will cause undefined behaviour. Creating this slice will also instatiate and Altrep objects.

unsafe fn as_typed_slice_raw_mut<T>(&mut self) -> &mut [T]

Get a mutable slice to this object’s data.

Safety

Unless the type is correct, this will cause undefined behaviour. Creating this slice will also instatiate and Altrep objects. Not all obejects (especially not list and strings) support this.

Object Safety

This trait is not object safe.