extendr_api::robj

Trait ToVectorValue

pub trait ToVectorValue {
    // Provided methods
    fn sexptype() -> SEXPTYPE { ... }
    fn to_real(&self) -> f64
       where Self: Sized { ... }
    fn to_complex(&self) -> Rcomplex
       where Self: Sized { ... }
    fn to_integer(&self) -> i32
       where Self: Sized { ... }
    fn to_logical(&self) -> i32
       where Self: Sized { ... }
    fn to_raw(&self) -> u8
       where Self: Sized { ... }
    fn to_sexp(&self) -> SEXP
       where Self: Sized { ... }
}

ToVectorValue is a trait that allows many different types to be converted to vectors. It is used as a type parameter to collect_robj().

Provided Methods

fn sexptype() -> SEXPTYPE

fn to_real(&self) -> f64
where Self: Sized,

fn to_complex(&self) -> Rcomplex
where Self: Sized,

fn to_integer(&self) -> i32
where Self: Sized,

fn to_logical(&self) -> i32
where Self: Sized,

fn to_raw(&self) -> u8
where Self: Sized,

fn to_sexp(&self) -> SEXP
where Self: Sized,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.