extendr_api::thread_safety

Function register_extendr_panic_hook

#[unsafe(no_mangle)]
pub extern "C" fn register_extendr_panic_hook()

This function registers a configurable print panic hook, for use in extendr-based R-packages. If the environment variable EXTENDR_BACKTRACE is set to either true or 1, then it displays the entire Rust panic traceback (default hook), otherwise it omits the panic backtrace.