extendr_api

Trait TryInto
·

pub trait TryInto<T>: Sized {
    type Error;

    // Required method
    fn try_into(self) -> Result<T, Self::Error>;
}

An attempted conversion that consumes self, which may or may not be expensive.

Library authors should usually not directly implement this trait, but should prefer implementing the TryFrom trait, which offers greater flexibility and provides an equivalent TryInto implementation for free, thanks to a blanket implementation in the standard library. For more information on this, see the documentation for Into.

Prefer using TryInto over TryFrom when specifying trait bounds on a generic function to ensure that types that only implement TryInto can be used as well.

Implementing TryInto

This suffers the same restrictions and reasoning as implementing Into, see there for details.

Required Associated Types

type Error

The type returned in the event of a conversion error.

Required Methods

fn try_into(self) -> Result<T, Self::Error>

Performs the conversion.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.