Module extendr_api::wrapper

Wrappers are lightweight proxies for references to R datatypes. They do not contain an Robj (see array.rs for an example of this).

Re-exports

pub use character::Character;

pub use environment::EnvIter;

pub use environment::Environment;

pub use expr::Expression;

pub use function::Function;

pub use lang::Language;

pub use list::FromList;

pub use list::List;

pub use list::ListIter;

pub use matrix::RArray;

pub use matrix::RColumn;

pub use matrix::RMatrix;

pub use matrix::RMatrix3D;

pub use nullable::Nullable;

pub use pairlist::Pairlist;

pub use pairlist::PairlistIter;

pub use primitive::Primitive;

pub use promise::Promise;

pub use raw::Raw;

pub use symbol::Symbol;

Modules

character

environment

expr

function

lang

list

matrix
Wrappers for matrices with deferred arithmetic.

nullable

pairlist
A pairlist is a linked list of values with optional symbol tags.

primitive

promise

raw

symbol

Traits

SymPair