Module extendr_api::wrapper::altrep

Macros

make_from_iterator ğŸ”’

Structs

Altrep

Traits

AltComplexImpl

AltIntegerImpl

AltLogicalImpl

AltRawImpl

AltRealImpl

AltStringImpl

AltrepImpl

Rust trait for implementing ALTREP. Implement one or more of these methods to generate an Altrep class. This is likely to be unstable for a while.

Functions

altcomplex_Elt ğŸ”’ âš

altcomplex_Get_region ğŸ”’ âš

altinteger_Elt ğŸ”’ âš

altinteger_Get_region ğŸ”’ âš

altinteger_Is_sorted ğŸ”’ âš

altinteger_Max ğŸ”’ âš

altinteger_Min ğŸ”’ âš

altinteger_No_NA ğŸ”’ âš

altinteger_Sum ğŸ”’ âš

altlogical_Elt ğŸ”’ âš

altlogical_Get_region ğŸ”’ âš

altlogical_Is_sorted ğŸ”’ âš

altlogical_No_NA ğŸ”’ âš

altlogical_Sum ğŸ”’ âš

altraw_Elt ğŸ”’ âš

altraw_Get_region ğŸ”’ âš

altreal_Elt ğŸ”’ âš

altreal_Get_region ğŸ”’ âš

altreal_Is_sorted ğŸ”’ âš

altreal_Max ğŸ”’ âš

altreal_Min ğŸ”’ âš

altreal_No_NA ğŸ”’ âš

altreal_Sum ğŸ”’ âš

altrep_Coerce ğŸ”’ âš

altrep_Duplicate ğŸ”’ âš

altrep_DuplicateEX ğŸ”’ âš

altrep_Inspect ğŸ”’ âš

altrep_Length ğŸ”’ âš

altrep_Serialized_state ğŸ”’ âš

altrep_Unserialize ğŸ”’ âš

altrep_UnserializeEX ğŸ”’ âš

altstring_Elt ğŸ”’ âš

altstring_Is_sorted ğŸ”’ âš

altstring_No_NA ğŸ”’ âš

altstring_Set_elt ğŸ”’ âš

altvec_Dataptr ğŸ”’ âš

altvec_Dataptr_or_null ğŸ”’ âš

altvec_Extract_subset ğŸ”’ âš

finalizer ğŸ”’ âš

manifest ğŸ”’