Struct extendr_api::wrapper::altrep::Altrep

pub struct Altrep { /* private fields */ }

Implementations

impl Altrep

pub fn data(&self) -> (Robj, Robj)

Safely implement R_altrep_data1, R_altrep_data2. When implementing Altrep classes, this gets the metadata.

pub fn set_data(&mut self, values: (Robj, Robj))

Safely (relatively!) implement R_set_altrep_data1, R_set_altrep_data2. When implementing Altrep classes, this sets the metadata.

pub fn class(&self) -> Robj

Safely implement ALTREP_CLASS.

pub fn from_state_and_class<StateType: 'static>( state: StateType, class: Robj, mutable: bool ) -> Altrep

pub fn is_manifest(&self) -> bool

Return true if the ALTREP object has been manifested (copied into memory).

pub fn make_altinteger_class<StateType: AltrepImpl + AltIntegerImpl + 'static>( name: &str, base: &str ) -> Robj

Make an integer ALTREP class that can be used to make vectors.

pub fn make_altreal_class<StateType: AltrepImpl + AltRealImpl + 'static>( name: &str, base: &str ) -> Robj

Make a real ALTREP class that can be used to make vectors.

pub fn make_altlogical_class<StateType: AltrepImpl + AltLogicalImpl + 'static>( name: &str, base: &str ) -> Robj

Make a logical ALTREP class that can be used to make vectors.

pub fn make_altraw_class<StateType: AltrepImpl + AltRawImpl + 'static>( name: &str, base: &str ) -> Robj

Make a raw ALTREP class that can be used to make vectors.

pub fn make_altcomplex_class<StateType: AltrepImpl + AltComplexImpl + 'static>( name: &str, base: &str ) -> Robj

Make a complex ALTREP class that can be used to make vectors.

pub fn make_altstring_class<StateType: AltrepImpl + AltStringImpl + 'static>( name: &str, base: &str ) -> Robj

Make a complex ALTREP class that can be used to make vectors.

pub fn make_altinteger_from_iterator<Iter>(iter: Iter) -> Altrepwhere Iter: ExactSizeIterator + Debug + Clone + 'static + Any, Iter::Item: Into<Rint>,

pub fn make_altlogical_from_iterator<Iter>(iter: Iter) -> Altrepwhere Iter: ExactSizeIterator + Debug + Clone + 'static + Any, Iter::Item: Into<Rbool>,

pub fn make_altreal_from_iterator<Iter>(iter: Iter) -> Altrepwhere Iter: ExactSizeIterator + Debug + Clone + 'static + Any, Iter::Item: Into<Rfloat>,

pub fn make_altcomplex_from_iterator<Iter>(iter: Iter) -> Altrepwhere Iter: ExactSizeIterator + Debug + Clone + 'static + Any, Iter::Item: Into<Rcplx>,