Trait extendr_api::wrapper::altrep::AltIntegerImpl

pub trait AltIntegerImpl: AltrepImpl {
    fn elt(&self, _index: usize) -> i32;

    fn tot_min_max_nas(&self) -> (i64, i32, i32, usize, usize) { ... }

    fn get_region(&self, index: usize, data: &mut [i32]) -> usize { ... }

    fn is_sorted(&self) -> Bool { ... }

    fn no_na(&self) -> bool { ... }

    fn sum(&self, remove_nas: bool) -> Robj { ... }

    fn min(&self, remove_nas: bool) -> Robj { ... }

    fn max(&self, remove_nas: bool) -> Robj { ... }
}

Required methods

fn elt(&self, _index: usize) -> i32

Get a single element from this vector.

Provided methods

fn tot_min_max_nas(&self) -> (i64, i32, i32, usize, usize)

fn get_region(&self, index: usize, data: &mut [i32]) -> usize

Get a multiple elements from this vector.

fn is_sorted(&self) -> Bool

Return TRUE if this vector is sorted, FALSE if not and NA_LOGICAL if unknown.

fn no_na(&self) -> bool

Return true if this vector does not contain NAs.

fn sum(&self, remove_nas: bool) -> Robj

Return the sum of the elements in this vector. If remove_nas is true, skip and NA values.

fn min(&self, remove_nas: bool) -> Robj

Return the minimum of the elements in this vector. If remove_nas is true, skip and NA values.

fn max(&self, remove_nas: bool) -> Robj

Return the maximum of the elements in this vector. If remove_nas is true, skip and NA values.