extendr_api::wrapper::altrep

Trait AltListImpl

pub trait AltListImpl {
    // Required method
    fn elt(&self, _index: usize) -> Robj;

    // Provided method
    fn set_elt(&mut self, _index: usize, _value: Robj) { ... }
}

Required Methods

fn elt(&self, _index: usize) -> Robj

Get a single element from this vector a single element of a list can be any Robj

Provided Methods

fn set_elt(&mut self, _index: usize, _value: Robj)

Set a single element in this list.