extendr_api::wrapper::altrep

Trait AltLogicalImpl

pub trait AltLogicalImpl: AltrepImpl {
    // Required method
    fn elt(&self, _index: usize) -> Rbool;

    // Provided methods
    fn tot_min_max_nas(&self) -> (i64, i32, i32, usize, usize) { ... }
    fn get_region(&self, index: usize, data: &mut [Rbool]) -> usize { ... }
    fn is_sorted(&self) -> Rbool { ... }
    fn no_na(&self) -> bool { ... }
    fn sum(&self, remove_nas: bool) -> Robj { ... }
}

Required Methods

fn elt(&self, _index: usize) -> Rbool

Get a single element from this vector.

Provided Methods

fn tot_min_max_nas(&self) -> (i64, i32, i32, usize, usize)

fn get_region(&self, index: usize, data: &mut [Rbool]) -> usize

Get a multiple elements from this vector.

fn is_sorted(&self) -> Rbool

Return TRUE if this vector is sorted, FALSE if not and Rbool::na() if unknown.

fn no_na(&self) -> bool

Return true if this vector does not contain NAs.

fn sum(&self, remove_nas: bool) -> Robj

Return the sum of the elements in this vector. If remove_nas is true, skip and NA values.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.