extendr_api::wrapper

Module doubles

Structs

Doubles

An obscure NA-aware wrapper for R’s double vectors. Can be used to iterate over vectors obtained from R or to create new vectors that can be returned back to R.