extendr_api::wrapper::matrix

Struct RArray

pub struct RArray<T, const NDIM: usize> {
    robj: Robj,
    _data: PhantomData<T>,
}

Wrapper for creating and using matrices and arrays.

use extendr_api::prelude::*;
test! {
    let matrix = RMatrix::new_matrix(3, 2, |r, c| [
        [1., 2., 3.],
         [4., 5., 6.]][c][r]);
    let robj = r!(matrix);
    assert_eq!(robj.is_matrix(), true);
    assert_eq!(robj.nrows(), 3);
    assert_eq!(robj.ncols(), 2);

    let matrix2 : RMatrix<f64> = robj.as_matrix().ok_or("error")?;
    assert_eq!(matrix2.data().len(), 6);
    assert_eq!(matrix2.nrows(), 3);
    assert_eq!(matrix2.ncols(), 2);
}

Fields
robj: Robj

Owning Robj (probably should be a Pin).

_data: PhantomData<T>

Data type of the n-array

Implementations

impl<T, const NDIM: usize> RArray<T, NDIM>

pub fn get_dimnames(&self) -> List

pub fn get_dim(&self) -> Vec<usize>

Get the dimension vector of the array.

Equivalent to dim() in R

pub fn set_names(&mut self, names: Strings)

Set the names of the elements of an array.

Equivalent to names<- in R

pub fn set_dimnames(&mut self, dimnames: List)

Set the dimension names of an array.

For RMatrix a list of length 2 is required, as that would entail column-names and row-names. If you only wish to set one, but not the other, then the unset element must be R NULL

Equivalent to dimnames<- in R

pub fn set_dim(&mut self, dim: Robj)

Set the dimensions of an array.

Equivalent to dim<-

impl<T, const NDIM: usize> RArray<T, NDIM>
where T: ToVectorValue, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new_array(dim: [usize; NDIM]) -> Self

impl<T> RArray<T, 2>
where T: ToVectorValue, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new(nrow: usize, ncol: usize) -> Self

Returns an RMatrix with dimensions according to nrow and ncol, with arbitrary entries. To initialize a matrix containing only NA values, use RMatrix::new_with_na.

impl<T> RArray<T, 2>
where T: ToVectorValue + CanBeNA, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new_with_na(nrow: usize, ncol: usize) -> Self

Returns an RMatrix with dimensions according to nrow and ncol, with all entries set to NA.

Note that since Raw does not have an NA representation in R, this method is not implemented for [Rbyte].

impl<T> RArray<T, 2>

pub fn get_colnames(&self) -> Option<Strings>

pub fn get_rownames(&self) -> Option<Strings>

impl<T, const NDIM: usize> RArray<T, NDIM>
where Robj: for<'a> AsTypedSlice<'a, T>,

pub fn from_parts(robj: Robj) -> Self

pub fn data(&self) -> &[T]

Returns a flat representation of the array in col-major.

pub fn data_mut(&mut self) -> &mut [T]

Returns a flat, mutable representation of the array in col-major.

pub fn ndim(&self) -> usize

Returns the number of dimensions.

pub fn dim(&self) -> Vec<usize>

Returns the dimensions of the array.

impl<T> RArray<T, 1>
where T: ToVectorValue, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new_column<F: FnMut(usize) -> T>(nrows: usize, f: F) -> Self

Make a new column type.

pub fn nrows(&self) -> usize

Get the number of rows.

impl<T> RArray<T, 2>
where T: ToVectorValue, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new_matrix<F: Clone + FnMut(usize, usize) -> T>( nrows: usize, ncols: usize, f: F, ) -> Self

Create a new matrix wrapper.

Arguments

nrows - the number of rows the returned matrix will have

ncols - the number of columns the returned matrix will have

f - a function that will be called for each entry of the matrix in order to populate it with values. It must return a scalar value that can be converted to an R scalar, such as i32, u32, or f64, i.e. see ToVectorValue. It accepts two arguments:

r - the current row of the entry we are creating

c - the current column of the entry we are creating

pub fn nrows(&self) -> usize

Get the number of rows.

pub fn ncols(&self) -> usize

Get the number of columns.

impl<T> RArray<T, 3>
where T: ToVectorValue, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new_matrix3d<F: Clone + FnMut(usize, usize, usize) -> T>( nrows: usize, ncols: usize, nmatrix: usize, f: F, ) -> Self

pub fn nrows(&self) -> usize

Get the number of rows.

pub fn ncols(&self) -> usize

Get the number of columns.

pub fn nsub(&self) -> usize

Get the number of submatrices.