Crate extendr_api

extendr - A safe and user friendly R extension interface.

This library aims to provide an interface that will be familiar to first-time users of Rust or indeed any compiled language.

Anyone who knows the R library should be able to write R extensions.

See the Robj struct for much of the content of this crate. Robj provides a safe wrapper for the R object type.

This library is just being born, but goals are:

Implement common R functions such as c() and print()

Example:

let v = c!(1, 2, 3);
let l = list!(a=1, b=2);
print!(v, l);

Provide a wrapper for r objects.

Example:

let s = Robj::from("hello");
let i = Robj::from(1);
let r = Robj::from(1.0);

Provide iterator support for creation and consumption of r vectors.

Example:

let res = (1..=100).iter().collect::<Robj>();
for x in res {
    print!(x);
}

Provide a procedural macro to adapt Rust functions to R

Example:

#[extendr]
fn fred(a: i32) -> i32 {
    a + 1
}

In R:

result <- .Call("fred", 1)

Macros

append_lang
Convert a list of tokens to an array of tuples.

args

c
Concatenation operator.

data_frame
Create a dataframe.

extendr_module
Define a module and export symbols to R Example:

factor
Create a factor.

lang
A macro for constructing R langage objects.

list
Create a list.

push_args
Convert a list of tokens to an array of tuples.

r
R object encapsulation operator.

read_table
Read a CSV file.

rep
Create a vector with repeating elements.

reprint
Print via the R error stream.

reprintln
Print with a newline via the R output stream.

rprint
Print via the R output stream.

rprintln
Print with a newline via the R output stream.

Structs

CallMethod

Character
Wrapper for creating character objects.

Lang
Wrapper for creating language objects.

List
Wrapper for creating list objects.

ListIter

StrIter

Symbol
Wrapper for creating symbols.

VecIter

Enums

Robj
Wrapper for an R S-expression pointer (SEXP).

Constants

FALSE

NULL

TRUE

Traits

AsTypedSlice

FromRobj
Trait used for incomming parameter conversion.

RobjItertools

ToVectorValue
ToVectorValue is a trait that allows many different types to be converted to vectors. It is used as a type parameter to collect_robj().

Functions

append⚠

append_with_name⚠

handle_panic
This function is use by the wrapper logic to catch panics on return.

integer_is_na

make_lang⚠

new_borrowed⚠

new_owned⚠

new_sys⚠

numeric_is_na

print_r_error

print_r_output

register_call_methods⚠

unwrap_or_throw
Throw an R error if a result is an error.

Type Definitions

AnyError
Generic dynamic error type.

DllInfo

IntegerIter
Iterator over the objects in a vector or string.

LogicalIter

NumericIter

SEXP

Attribute Macros

extendr
Generate bindings for a single function.