Enum extendr_api::Robj

pub enum Robj {
    Owned(SEXP),
    Borrowed(SEXP),
    Sys(SEXP),
}

Wrapper for an R S-expression pointer (SEXP).

As much as possible we wish to make this object safe (ie. no segfaults).

If you avoid using unsafe functions it is more likely that you will avoid panics and segfaults. We will take great trouble to ensure that this is true.

Variants

Owned(SEXP)

This object owns the SEXP and must free it.

Borrowed(SEXP)

This object references a SEXP such as an input parameter.

Sys(SEXP)

This object references a SEXP owned by libR.

Implementations

impl Robj

pub unsafe fn get(&self) -> SEXP

Get a copy of the underlying SEXP. Note: this is unsafe.

pub unsafe fn get_mut(&mut self) -> Option<SEXP>

Get a copy of the underlying SEXP for mutable types. This is valid only for owned objects as we are not permitted to modify parameters or system objects.

pub fn sexptype(&self) -> u32

Get the XXXSXP type of the object.

pub fn len(&self) -> usize

Get the extended length of the object.

pub fn as_i32_slice(&self) -> Option<&[i32]>

Get a read-only reference to the content of an integer or logical vector.

pub fn as_integer_slice(&self) -> Option<&[i32]>

Get a read-only reference to the content of an integer or logical vector.

pub fn as_integer_vector(&self) -> Option<Vec<i32>>

Get a Vec copied from the object.

pub fn as_logical_slice(&self) -> Option<&[Bool]>

Get a read-only reference to the content of an integer or logical vector.

pub fn as_logical_vector(&self) -> Option<Vec<bool>>

Get a Vec copied from the object.

pub fn as_f64_slice(&self) -> Option<&[f64]>

Get a read-only reference to the content of a double vector.

pub fn as_numeric_slice(&self) -> Option<&[f64]>

Get a read-only reference to the content of a double vector.

pub fn as_numeric_vector(&self) -> Option<Vec<f64>>

Get a Vec copied from the object.

pub fn as_u8_slice(&self) -> Option<&[u8]>

Get a read-only reference to the content of an integer or logical vector.

pub fn as_raw_slice(&self) -> Option<&[u8]>

Get a read-only reference to the content of an integer or logical vector.

pub fn as_i32_slice_mut(&mut self) -> Option<&mut [i32]>

Get a read-write reference to the content of an integer or logical vector.

pub fn as_f64_slice_mut(&mut self) -> Option<&mut [f64]>

Get a read-write reference to the content of a double vector.

pub fn as_u8_slice_mut(&mut self) -> Option<&mut [u8]>

Get a read-write reference to the content of an integer or logical vector.

pub fn pairlist_iter(&self) -> Option<ListIter>

Get an iterator over a pairlist.

pub fn list_iter(&self) -> Option<VecIter>

Get an iterator over an unnamed list.

pub fn str_iter(&self) -> Option<StrIter>

Get an iterator over a string vector.

pub fn as_str(&self) -> Option<&str>

Get a read-only reference to a char, symbol or string type.

pub fn as_i32(&self) -> Option<i32>

Get a scalar integer.

pub fn as_integer(&self) -> Option<i32>

Get a scalar integer.

pub fn as_f64(&self) -> Option<f64>

Get a scalar real.

pub fn as_numeric(&self) -> Option<f64>

Get a scalar real.

pub fn as_bool(&self) -> Option<bool>

Get a scalar boolean. Also accepts 1/0.

pub fn eval(&self) -> Result<Robj, AnyError>

Evaluate the expression and return an error or an R object.

pub fn eval_blind(&self) -> Robj

Evaluate the expression and return NULL or an R object.

pub fn parse(code: &str) -> Result<Robj, AnyError>

Parse a string into an R executable object

pub fn eval_string(code: &str) -> Result<Robj, AnyError>

Parse a string into an R executable object and run it.

pub unsafe fn unprotected(self) -> Robj

Unprotect an object - assumes a transfer of ownership. This is unsafe because the object pointer may be left dangling.

pub fn is_owned(&self) -> bool

Return true if the object is owned by this wrapper. If so, it will be released when the wrapper drops.

impl Robj

pub fn names(&self) -> Option<StrIter>

Get the names attribute as a string iterator if one exists.

pub fn namesAndValues(&self) -> Option<Zip<StrIter, VecIter>>

Return an iterator over names and values of a list if they exist.

impl Robj

The following impls wrap specific Rinternals.h symbols.

pub fn globalEnv() -> Robj

The "global" environment

pub fn emptyEnv() -> Robj

An empty environment at the root of the environment tree

pub fn baseEnv() -> Robj

The base environment; formerly R_NilValue

pub fn baseNamespace() -> Robj

The (fake) namespace for base

pub fn namespaceRegistry() -> Robj

for registered namespaces

pub fn srcref() -> Robj

Current srcref, for debuggers

pub fn nilValue() -> Robj

The nil object

pub fn unboundValue() -> Robj

Unbound marker

pub fn missingArg() -> Robj

Missing argument marker

pub fn baseSymbol() -> Robj

"base"

pub fn braceSymbol() -> Robj

"{"

pub fn bracket2Symbol() -> Robj

"[["

pub fn bracketSymbol() -> Robj

"["

pub fn classSymbol() -> Robj

"class"

pub fn deviceSymbol() -> Robj

".Device"

pub fn dimNamesSymbol() -> Robj

"dimnames"

pub fn dimSymbol() -> Robj

"dim"

pub fn dollarSymbol() -> Robj

"$"

pub fn dotsSymbol() -> Robj

"..."

pub fn doubleColonSymbol() -> Robj

pub fn dropSymbol() -> Robj { unsafe { new_sys(R_DropSymbol) }}"drop"

pub fn lastvalueSymbol() -> Robj

".Last.value"

pub fn levelsSymbol() -> Robj

"levels"

pub fn modeSymbol() -> Robj

"mode"

pub fn naRmSymbol() -> Robj

"na.rm"

pub fn nameSymbol() -> Robj

"name"

pub fn namesSymbol() -> Robj

"names"

pub fn namespaceEnvSymbol() -> Robj

NAMESPACE_."

pub fn packageSymbol() -> Robj

"package"

pub fn previousSymbol() -> Robj

"previous"

pub fn quoteSymbol() -> Robj

"quote"

pub fn rowNamesSymbol() -> Robj

"row.names"

pub fn seedsSymbol() -> Robj

".Random.seed"

pub fn sortListSymbol() -> Robj

"sort.list"

pub fn sourceSymbol() -> Robj

"source"

pub fn specSymbol() -> Robj

"spec"

pub fn tspSymbol() -> Robj

"tsp"

pub fn tripleColonSymbol() -> Robj

":::"

pub fn dot_defined() -> Robj

".defined"

pub fn dot_Method() -> Robj

".Method"

pub fn dot_packageName() -> Robj

"packageName"

pub fn dot_target() -> Robj

".target"

pub fn naString() -> Robj

NA_STRING as a CHARSXP

pub fn blankString() -> Robj

"" as a CHARSXP

pub fn blankScalarString() -> Robj

as a STRSXP

impl Robj

The following impls wrap specific Rinternals.h functions.

pub fn isNull(&self) -> bool

Return true if this is the null object.

pub fn isSymbol(&self) -> bool

Return true if this is a symbol.

pub fn isLogical(&self) -> bool

Return true if this is a boolean (logical) vector

pub fn isReal(&self) -> bool

Return true if this is a real (f64) vector.

pub fn isComplex(&self) -> bool

Return true if this is a complex vector.

pub fn isExpression(&self) -> bool

Return true if this is an expression.

pub fn isEnvironment(&self) -> bool

Return true if this is an environment.

pub fn isString(&self) -> bool

Return true if this is a string.

pub fn isObject(&self) -> bool

Return true if this is an object.

pub fn getCurrentSrcref(val: i32) -> Robj

Get the source ref.

pub fn getSrcFilename(&self) -> Robj

Get the source filename.

pub fn asChar(&self) -> Robj

Convert to a string vector.

pub fn coerceVector(&self, sexptype: u32) -> Robj

Convert to vectors of many kinds.

pub fn pairToVectorList(&self) -> Robj

Convert a pairlist (LISTSXP) to a vector list (VECSXP).

pub fn vectorToPairList(&self) -> Robj

Convert a vector list (VECSXP) to a pair list (LISTSXP)

pub fn asCharacterFactor(&self) -> Robj

Assign an integer to each unique string and return a "factor".

pub fn asLogical(&self) -> bool

Get a scalar boolean value

pub fn asInteger(&self) -> i32

Get a scalar 32 bit integer value

pub fn asReal(&self) -> f64

Get a 64 bit double value

pub fn allocMatrix(sexptype: SEXPTYPE, rows: i32, cols: i32) -> Robj

Allocate a matrix object (see NumericMatrix etc.)

pub fn duplicate(&self) -> Self

Compatible way to duplicate an object. Use obj.clone() instead for Rust compatibility.

pub fn getAttrib(&self, name: &Robj) -> Robj

Get a specific attribute as a borrowed robj.

pub fn ncols(&self) -> usize

Number of columns of a matrix

pub fn nrows(&self) -> usize

Number of rows of a matrix

pub unsafe fn makeExternalPtr<T>(p: *mut T, tag: Robj, prot: Robj) -> Self

Internal function used to implement #[extendr] impl

pub unsafe fn externalPtrAddr<T>(&self) -> *mut T

Internal function used to implement #[extendr] impl

pub unsafe fn externalPtrTag(&self) -> Self

Internal function used to implement #[extendr] impl

pub unsafe fn externalPtrProtected(&self) -> Self

Internal function used to implement #[extendr] impl

pub unsafe fn registerCFinalizer(&self, func: R_CFinalizer_t)

pub fn xlengthgets(&self, new_len: usize) -> Result<Robj, AnyError>

Read-only access to attribute list. Copy a vector and resize it. See. https://github.com/hadley/r-internals/blob/master/vectors.md

pub fn allocVector(sexptype: u32, len: usize) -> Robj

Allocated an owned object of a certain type.

pub fn conformable(a: &Robj, b: &Robj) -> bool

Return true if two arrays have identical dims.

pub fn elt(&self, index: usize) -> Robj

Borrow an element from a list.

pub fn isArray(&self) -> bool

Return true if this is an array.

pub fn isFactor(&self) -> bool

Return true if this is factor.

pub fn isFrame(&self) -> bool

Return true if this is a data frame.

pub fn isFunction(&self) -> bool

Return true if this is a function.

pub fn isInteger(&self) -> bool

Return true if this is an integer vector.

pub fn isLanguage(&self) -> bool

Return true if this is a language object.

pub fn isList(&self) -> bool

Return true if this is a vector list.

pub fn isMatrix(&self) -> bool

Return true if this is a matrix.

pub fn isNewList(&self) -> bool

Return true if this is a vector list or null.

pub fn isNumber(&self) -> bool

Return true if this is a numeric vector but not a factor.

pub fn isNumeric(&self) -> bool

Return true if this is a numeric vector but not a factor or complex.

pub fn isPairList(&self) -> bool

Return true if this is a pairlist.

pub fn isPrimitive(&self) -> bool

Return true if this is a primitive function.

pub fn isTs(&self) -> bool

Return true if this is a time series vector (see tsp).

pub fn isUserBinop(&self) -> bool

Return true if this is a user defined binop.

pub fn isValidString(&self) -> bool

Return true if this is a valid string.

pub fn isValidStringF(&self) -> bool

Return true if this is a valid string.

pub fn isVector(&self) -> bool

Return true if this is a vector.

pub fn isVectorAtomic(&self) -> bool

Return true if this is an atomic vector.

pub fn isVectorList(&self) -> bool

Return true if this is a vector list.

pub fn isVectorizable(&self) -> bool

Return true if this is can be made into a vector.

pub fn check_external_ptr(&self, expected_tag: &str) -> bool

Check an external pointer tag This may work better by using a symbol cached in a static variable.