Module extendr_api::functions

Functions

base_env

The base environment; formerly R_NilValue

base_namespace

The namespace for base.

blank_scalar_string

“” as a STRSXP

blank_string

“” as a CHARSXP

current_env

The current interpreter environment.

empty_env

An empty environment at the root of the environment tree

eval_string

Parse a string into an R executable object and run it. Used by the R! macro.

eval_string_with_params

Parse a string into an R executable object and run it using parameters param.0, param.1, …

find_namespace

Find a namespace by name.

find_namespaced_function

Find a function or primitive that may be in a namespace.

global_env

The “global” environment

global_function

Get a global function from global_env() and ancestors.

global_var

Get a global variable from global_env() and ancestors. If the result is a promise, evaulate the promise.

local_var

Get a local variable from current_env() and ancestors.

na_string

NA_STRING as a CHARSXP

namespace_registry

For registered namespaces.

new_env

Create a new environment

nil_value

The nil object

parse

Parse a string into an R executable object

srcref

Current srcref, for debuggers