Module extendr_api::ownership

Maintain ownership of R objects.

This provides the functions protect and unprotect. A single preserved vector holds ownership of all protected objects.

Objects are reference counted, so multiple calls are possible, unlike R_PreserveObject.

This module exports two functions, protect(sexp) and unprotect(sexp).

Structs

Object 🔒

Ownership 🔒

Constants

EXTRA_PRESERVATION_SIZE

INITIAL_PRESERVATION_SIZE

Statics

OWNERSHIP 🔒

Functions

protect 🔒 ⚠

unprotect 🔒 ⚠