Function extendr_api::prelude::stack

pub fn stack<A, D>(
    axis: Axis, 
    arrays: &[ArrayBase<ViewRepr<&'a A>, D>]
) -> Result<ArrayBase<OwnedRepr<A>, D>, ShapeError> where
    A: Copy,
    D: RemoveAxis, 

Stack arrays along the given axis.

Errors if the arrays have mismatching shapes, apart from along axis. (may be made more flexible in the future).
Errors if arrays is empty, if axis is out of bounds, if the result is larger than is possible to represent.

use ndarray::{arr2, Axis, stack};

let a = arr2(&[[2., 2.],
               [3., 3.]]);
assert!(
    stack(Axis(0), &[a.view(), a.view()])
    == Ok(arr2(&[[2., 2.],
                 [3., 3.],
                 [2., 2.],
                 [3., 3.]]))
);