extendr_api::prelude::modules::cholesky::llt::compute
Function cholesky_in_place

pub fn cholesky_in_place<E>(
    matrix: MatMut<'_, E>,
    regularization: LltRegularization<E>,
    parallelism: Parallelism<'_>,
    stack: PodStack<'_>,
    params: LltParams,
) -> Result<LltInfo, CholeskyError>
where
    E: ComplexField,

Computes the Cholesky factor $L$ of a Hermitian positive definite input matrix $A$ such that $L$ is lower triangular, and $$LL^H == A.$$

The result is stored back in the lower half of the same matrix, or an error is returned if the matrix is not positive definite.

The input matrix is interpreted as Hermitian with the values being extracted from the lower part, but the entire matrix is required to be initialized.

The strictly upper triangular part of the matrix is clobbered and may be filled with garbage values.

Panics

Panics if the input matrix is not square.

This can also panic if the provided memory in stack is insufficient (see cholesky_in_place_req).