Module extendr_api::prelude::modules::core::inverse

Triangular matrix inversion.

Functions

invert_lower_triangular

Computes the inverse of the lower triangular matrix src and stores the lower triangular part of the result to dst.

invert_unit_lower_triangular

Computes the inverse of the lower triangular matrix src (with implicit unit diagonal) and stores the strictly lower triangular part of the result to dst.

invert_unit_upper_triangular

Computes the inverse of the upper triangular matrix src (with implicit unit diagonal) and stores the strictly upper triangular part of the result to dst.

invert_upper_triangular

Computes the inverse of the upper triangular matrix src and stores the upper triangular part of the result to dst.