Module extendr_api::prelude::modules::core::zip

Implementation of zipped! structures.

Structs

Last

Single element.

LastEq

Single matrix view.

Read

Read only view over a single matrix element.

ReadWrite

Read-write view over a single matrix element.

Zip

Zipped elements.

ZipEq

Zipped matrix views.

Enums

Diag

Specifies whether the main diagonal should be traversed, when iterating over a triangular chunk of the matrix.

MatLayoutTransform

Matrix layout transformation. Used for zipping optimizations.

VecLayoutTransform

Vector layout transformation. Used for zipping optimizations.

Traits

MatIndex

Zipped matrix views.

MatShape

Type with a given matrix shape.

MaybeContiguous

Zipped matrix views.

ViewMut

Type that can be converted to a view.