Module extendr_api::prelude::prelude

ndarray prelude.

This module contains the most used types, type aliases, traits, functions, and macros that you can import easily as a group.

use ndarray::prelude::*;

Macros

array

Create an Array with one, two or three dimensions.

azip

Array zip macro: lock step function application across several arrays and producers.

s

Slice argument constructor.

Structs

ArrayBase

An n-dimensional array.

Axis

An axis index.

Dim

Dimension description.

NewAxis

Token to represent a new axis in a slice description.

Traits

AsArray

Argument conversion into an array view

Dimension

Array shape and index trait.

NdFloat

Floating-point element types f32 and f64.

ShapeBuilder

A trait for Shape and D where D: Dimension that allows customizing the memory layout (strides) of an array shape.

Functions

Dim

Create a new dimension value.

Ix0

Create a zero-dimensional index

Ix1

Create a one-dimensional index

Ix2

Create a two-dimensional index

Ix3

Create a three-dimensional index

Ix4

Create a four-dimensional index

Ix5

Create a five-dimensional index

Ix6

Create a six-dimensional index

IxDyn

Create a dynamic-dimensional index

arr0

Create a zero-dimensional array with the element x.

arr1

Create a one-dimensional array with elements from xs.

arr2

Create a two-dimensional array with elements from xs.

aview0

Create a zero-dimensional array view borrowing x.

aview1

Create a one-dimensional array view with elements borrowing xs.

aview2

Create a two-dimensional array view with elements borrowing xs.

aview_mut1

Create a one-dimensional read-write array view with elements borrowing xs.

Type Aliases

ArcArray

An array where the data has shared ownership and is copy on write.

Array

An array that owns its data uniquely.

Array0

zero-dimensional array

Array1

one-dimensional array

Array2

two-dimensional array

Array3

three-dimensional array

Array4

four-dimensional array

Array5

five-dimensional array

Array6

six-dimensional array

ArrayD

dynamic-dimensional array

ArrayView

A read-only array view.

ArrayView0

zero-dimensional array view

ArrayView1

one-dimensional array view

ArrayView2

two-dimensional array view

ArrayView3

three-dimensional array view

ArrayView4

four-dimensional array view

ArrayView5

five-dimensional array view

ArrayView6

six-dimensional array view

ArrayViewD

dynamic-dimensional array view

ArrayViewMut

A read-write array view.

ArrayViewMut0

zero-dimensional read-write array view

ArrayViewMut1

one-dimensional read-write array view

ArrayViewMut2

two-dimensional read-write array view

ArrayViewMut3

three-dimensional read-write array view

ArrayViewMut4

four-dimensional read-write array view

ArrayViewMut5

five-dimensional read-write array view

ArrayViewMut6

six-dimensional read-write array view

ArrayViewMutD

dynamic-dimensional read-write array view

CowArray

An array with copy-on-write behavior.

Ix0

zero-dimensionial

Ix1

one-dimensional

Ix2

two-dimensional

Ix3

three-dimensional

Ix4

four-dimensional

Ix5

five-dimensional

Ix6

six-dimensional

IxDyn

dynamic-dimensional

RawArrayView

A read-only array view without a lifetime.

RawArrayViewMut

A mutable array view without a lifetime.