Module extendr_api::prelude::sparse::linalg

Sparse linear algebra module.
Contains low level routines and the implementation of their corresponding high level wrappers.

Modules

amd

Approximate minimum degree ordering.

cholesky

Computes the Cholesky decomposition (either LLT, LDLT, or Bunch-Kaufman) of a given sparse matrix. See crate::linalg::cholesky for more info.

colamd

Approximate minimum degree column ordering.

lu

Computes the LU decomposition of a given sparse matrix. See faer::linalg::lu for more info.

matmul

Sparse matrix multiplication.

qr

Computes the QR decomposition of a given sparse matrix. See crate::linalg::qr for more info.

solvers

High level sparse solvers.

triangular_solve

Solving sparse triangular linear systems with a dense right-hand-side.

Structs

SupernodalThreshold

Nonnegative threshold controlling when the supernodal factorization is used.

SymbolicSupernodalParams

Tuning parameters for the supernodal factorizations.

Enums

CholeskyError

Sparse Cholesky error.

FaerError

Errors that can occur in sparse algorithms.

LuError

Sparse LU error.

Traits

Index

Trait for unsigned integers that can be indexed with.

SignedIndex

Trait for signed integers corresponding to the ones satisfying Index.