Module extendr_api::prelude::sparse::linalg::cholesky::supernodal

Supernodal factorization module.

A supernodal factorization is one that processes the elements of the Cholesky factor of the input matrix by blocks, rather than single elements. This is more efficient if the Cholesky factor is somewhat dense.

Structs

SupernodalIntranodeBunchKaufmanRef

Cholesky Bunch-Kaufman factors containing both the symbolic and numeric representations.

SupernodalLdltRef

Cholesky LDLT factors containing both the symbolic and numeric representations.

SupernodalLltRef

Cholesky LLT factor containing both its symbolic and numeric representations.

SupernodeRef

A single supernode from the Cholesky factor.

SymbolicSupernodalCholesky

Cholesky factor structure containing its symbolic structure.

SymbolicSupernodeRef

Symbolic structure of a single supernode from the Cholesky factor.

Functions

factorize_supernodal_numeric_intranode_bunch_kaufman

Computes the numeric values of the Cholesky Bunch-Kaufman factors of the matrix A with intranodal pivoting, and stores them in L_values.

factorize_supernodal_numeric_intranode_bunch_kaufman_req

Returns the size and alignment of the workspace required to compute the numeric Cholesky Bunch-Kaufman factorization with intranodal pivoting of a matrix A with dimension n.

factorize_supernodal_numeric_ldlt

Computes the numeric values of the Cholesky LDLT factors of the matrix A, and stores them in L_values.

factorize_supernodal_numeric_ldlt_req

Returns the size and alignment of the workspace required to compute the numeric Cholesky LDLT factorization of a matrix A with dimension n.

factorize_supernodal_numeric_llt

Computes the numeric values of the Cholesky LLT factor of the matrix A, and stores them in L_values.

factorize_supernodal_numeric_llt_req

Returns the size and alignment of the workspace required to compute the numeric Cholesky LLT factorization of a matrix A with dimension n.

factorize_supernodal_symbolic

Computes the supernodal symbolic structure of the Cholesky factor of the matrix A.

factorize_supernodal_symbolic_cholesky_req

Returns the size and alignment of the workspace required to compute the symbolic supernodal factorization of a matrix of size n.