Module extendr_api::prelude::sparse::linalg::lu::supernodal

Supernodal factorization module.

A supernodal factorization is one that processes the elements of the LU factors of the input matrix by blocks, rather than by single elements. This is more efficient if the LU factors are somewhat dense.

Structs

SupernodalLu

LU factor structure containing the symbolic and numerical representations.

SymbolicSupernodalLu

LU factor structure containing the symbolic structure.

Functions

factorize_supernodal_numeric_lu

Computes the numeric values of the LU factors of the matrix A as well as the row pivoting permutation, and stores them in lu and row_perm/row_perm_inv.

factorize_supernodal_numeric_lu_req

Computes the size and alignment of the workspace required to perform a numeric LU factorization.

factorize_supernodal_symbolic_lu

Computes the symbolic structure of the LU factors of the matrix A.

factorize_supernodal_symbolic_lu_req

Computes the size and alignment of the workspace required to compute the symbolic LU factorization of a square matrix with size n.