Struct extendr_api::prelude::SliceInfo

pub struct SliceInfo<T, Din, Dout> where
    Din: Dimension,
    Dout: Dimension,  { /* private fields */ }

Represents all of the necessary information to perform a slice.

The type T is typically [SliceInfoElem; n], &[SliceInfoElem], or Vec<SliceInfoElem>. The type Din is the dimension of the array to be sliced, and Dout is the output dimension after calling .slice(). Note that if Din is a fixed dimension type (Ix0, Ix1, Ix2, etc.), the SliceInfo instance can still be used to slice an array with dimension IxDyn as long as the number of axes matches.

Implementations

impl<T, Din, Dout> SliceInfo<T, Din, Dout> where
T: AsRef<[SliceInfoElem]>,
Din: Dimension,
Dout: Dimension,

pub unsafe fn new(indices: T) -> Result<SliceInfo<T, Din, Dout>, ShapeError>

Returns a new SliceInfo instance.

Errors if Din or Dout is not consistent with indices.

For common types, a safe alternative is to use TryFrom instead.

Safety

The caller must ensure indices.as_ref() always returns the same value when called multiple times.

pub fn in_ndim(&self) -> usize

Returns the number of dimensions of the input array for .slice().

If Din is a fixed-size dimension type, then this is equivalent to Din::NDIM.unwrap(). Otherwise, the value is calculated by iterating over the SliceInfoElem elements.

pub fn out_ndim(&self) -> usize

Returns the number of dimensions after calling .slice() (including taking subviews).

If Dout is a fixed-size dimension type, then this is equivalent to Dout::NDIM.unwrap(). Otherwise, the value is calculated by iterating over the SliceInfoElem elements.