Trait extendr_api::prelude::ShapeArg

pub trait ShapeArg {
    type Dim: Dimension;
    fn into_shape_and_order(self) -> (Self::Dim, Option<Order>);
}

Array shape argument with optional order parameter

Shape or array dimension argument, with optional Order parameter.

This is an argument conversion trait that is used to accept an array shape and (optionally) an ordering argument.

See for example .to_shape().

Associated Types

type Dim: Dimension

Required methods

fn into_shape_and_order(self) -> (Self::Dim, Option<Order>)