Module extendr_api::prelude::utils

Various utilities for low level implementations in generic code.

Modules

constrained

Index and matrix types with compile time checks, whichh can replace bound checks at runtime.

simd

Simd operations for a specific type satisfying ComplexField.

slice

Slice types for entities.

thread

Utilities relating to threading and parallelism.

vec

Vector type for entities.