Struct extendr_api::prelude::utils::constrained::Array

#[repr(transparent)]
pub struct Array<'n, T>(Branded<'n, [T]>);

Array of length equal to the value tied to 'n.

Tuple Fields
0: Branded<'n, [T]>
Implementations

impl<'n, T> Array<'n, T>

pub fn from_ref<'a>(slice: &'a [T], size: Size<'n>) -> &'a Array<'n, T>

Returns a constrained array after checking that its length matches size.

pub fn from_mut<'a>(slice: &'a mut [T], size: Size<'n>) -> &'a mut Array<'n, T>

Returns a constrained array after checking that its length matches size.

pub fn as_ref(&self) -> &[T]

Returns the unconstrained slice.

pub fn as_mut<'a>(&mut self) -> &'a mut [T]

Returns the unconstrained slice.

pub fn len(&self) -> Size<'n>

Returns the length of self.