Module extendr_api::robj

R object handling.

See. Writing R Extensions

Fundamental principals:

Any function that can break the protection mechanism is unsafe.

Users should be able to do almost everything without using libR_sys.

The interface should be friendly to R users without Rust experience.

Re-exports

pub use iter::*;

Modules

debug 🔒

from_robj 🔒

into_robj 🔒

operators 🔒

rinternals 🔒

try_from_robj 🔒

Conversions to Robj

Macros

make_typed_slice 🔒

Structs

Robj

Wrapper for an R S-expression pointer (SEXP).

Traits

AsTypedSlice

Generic access to typed slices in an Robj.

Attributes

Provides access to the attributes of an R object.

Eval

FromRobj

Trait used for incomming parameter conversion.

GetSexp

IntoRobj

This is an extension trait to provide a convenience method into_robj().

Length

Operators

The following impls add operators to Robj.

Rinternals

The following impls wrap specific Rinternals.h functions.

RobjItertools

Extensions to iterators for R objects including RobjItertools::collect_robj().

Slices

ToVectorValue

ToVectorValue is a trait that allows many different types to be converted to vectors. It is used as a type parameter to collect_robj().

Types

Functions

to_str 🔒 ⚠