Trait extendr_api::robj::GetSexp

pub trait GetSexp {
    unsafe fn get(&self) -> SEXP;
    fn as_robj(&self) -> &Robj;
    fn as_robj_mut(&mut self) -> &mut Robj;
}

Required Methods

unsafe fn get(&self) -> SEXP

Get a copy of the underlying SEXP.

Safety

Access to a raw SEXP pointer can cause undefined behaviour and is not thread safe.

fn as_robj(&self) -> &Robj

Get a reference to a Robj for this type.

fn as_robj_mut(&mut self) -> &mut Robj

Get a mutable reference to a Robj for this type.