Trait extendr_api::robj::SexpAsPtr

pub(crate) trait SexpAsPtr: Sized {
    type CType;

    const R_TYPE_ID: u32;
    const TYPE_ERROR: fn(_: Robj) -> Error;

    // Provided methods
    unsafe fn as_mut_ptr(x: SEXP) -> *mut Self { ... }
    unsafe fn as_ptr(x: SEXP) -> *const Self { ... }
}

Trait for providing rust with a typed pointer to an R scalar/vector type.

R data are hidden behind SEXP opaque pointers with an identifier for the underlying r storage type. Thus this trait is only meaningful for SEXP-wrappers.

Marker for the R object modes / types represented as a value.

Required Associated Types

type CType

Corresponding rust type to the type in C (in R’s C-API)

Required Associated Constants

const R_TYPE_ID: u32

See Rtype for reference.

const TYPE_ERROR: fn(_: Robj) -> Error

Provided Methods

unsafe fn as_mut_ptr(x: SEXP) -> *mut Self

Returns a typed mutable pointer to the underlying value

unsafe fn as_ptr(x: SEXP) -> *const Self

Object Safety

This trait is not object safe.