Trait extendr_api::robj::ToVectorValue

pub trait ToVectorValue {
    // Provided methods
    fn sexptype() -> SEXPTYPE { ... }
    fn to_real(&self) -> f64
       where Self: Sized { ... }
    fn to_complex(&self) -> Rcomplex
       where Self: Sized { ... }
    fn to_integer(&self) -> i32
       where Self: Sized { ... }
    fn to_logical(&self) -> i32
       where Self: Sized { ... }
    fn to_raw(&self) -> u8
       where Self: Sized { ... }
    fn to_sexp(&self) -> SEXP
       where Self: Sized { ... }
}

ToVectorValue is a trait that allows many different types to be converted to vectors. It is used as a type parameter to collect_robj().

Provided Methods

fn sexptype() -> SEXPTYPE

fn to_real(&self) -> f64where Self: Sized,

fn to_complex(&self) -> Rcomplexwhere Self: Sized,

fn to_integer(&self) -> i32where Self: Sized,

fn to_logical(&self) -> i32where Self: Sized,

fn to_raw(&self) -> u8where Self: Sized,

fn to_sexp(&self) -> SEXPwhere Self: Sized,