extendr_api::scalar

Module macros

Macros

gen_binop 🔒

Generates an implementation of a binary operator Trait for a scalar type

gen_binopassign 🔒

Generates an implementation of a binary operate-assign Trait for a scalar type

gen_from_primitive 🔒

Generates an implementation of type conversion Traits from a primitive type

gen_partial_ord 🔒

gen_sum_iter 🔒

Generates an implementation of std::iter::Sum for a scalar type

gen_trait_impl 🔒

Generates an implementation of a number of Traits for the specified Type

gen_unop 🔒

Generates an implementation of a unary operator Trait for a scalar type