Struct extendr_api::scalar::Rbool

#[repr(transparent)]
pub struct Rbool(i32);

Rbool is a wrapper for i32 in the context of an R’s logical vector.

Rbool can have a value of 0, 1 or i32::MIN.

The value i32::MIN is used as NA.

Rbool has the same footprint as an i32 value allowing us to use it in zero copy slices.

Tuple Fields
0: i32
Implementations

impl Rbool

pub const fn true_value() -> Rbool

Return a true Rbool.

pub const fn false_value() -> Rbool

Return a false Rbool.

pub const fn na_value() -> Rbool

Return a NA Rbool.

pub fn is_true(&self) -> bool

Return true if this triboolean is true but not NA.

pub fn is_false(&self) -> bool

Return true if this triboolean is false but not NA.

pub fn to_bool(&self) -> bool

Convert this Rbool to a bool. Note NA will be true.

pub fn from_bool(val: bool) -> Self

Convert this construct a Rbool from a rust boolean.