Module extendr_api::thread_safety

Provide limited protection for multithreaded access to the R API.

Constants

THREAD_HAS_LOCK 🔒

Statics

R_API_LOCK 🔒

A global lock, that should represent the global lock on the R-API. It is not tied to an actual instance of R.

R_ERROR_BUF 🔒

Functions

catch_r_error

Wrap an R function such as Rf_findFunction and convert errors and panics into results.

single_threaded

Run f while ensuring that f runs in a single-threaded manner.

throw_r_error