Trait extendr_api::ToVectorValue

pub trait ToVectorValue {
    fn sexptype() -> SEXPTYPE { ... }

    fn to_numeric(&self) -> f64
    where
        Self: Sized,
    { ... }

    fn to_integer(&self) -> i32
    where
        Self: Sized,
    { ... }

    fn to_logical(&self) -> i32
    where
        Self: Sized,
    { ... }

    fn to_raw(&self) -> u8
    where
        Self: Sized,
    { ... }

    fn to_sexp(&self) -> SEXP
    where
        Self: Sized,
    { ... }
}

ToVectorValue is a trait that allows many different types to be converted to vectors. It is used as a type parameter to collect_robj().

Provided methods

fn sexptype() -> SEXPTYPE

fn to_numeric(&self) -> f64 where
Self: Sized,

fn to_integer(&self) -> i32 where
Self: Sized,

fn to_logical(&self) -> i32 where
Self: Sized,

fn to_raw(&self) -> u8 where
Self: Sized,

fn to_sexp(&self) -> SEXP where
Self: Sized,

Loading content...