Module extendr_api::wrapper

Wrappers are lightweight proxies for references to R datatypes. They do not contain an Robj (see array.rs for an example of this).

Re-exports

pub use self::rstr::Rstr;

pub use altrep::AltListImpl;

pub use altrep::AltComplexImpl;

pub use altrep::AltIntegerImpl;

pub use altrep::AltLogicalImpl;

pub use altrep::AltRawImpl;

pub use altrep::AltRealImpl;

pub use altrep::AltStringImpl;

pub use altrep::Altrep;

pub use altrep::AltrepImpl;

pub use complexes::Complexes;

pub use dataframe::Dataframe;

pub use dataframe::IntoDataFrameRow;

pub use doubles::Doubles;

pub use environment::EnvIter;

pub use environment::Environment;

pub use expr::Expressions;

pub use externalptr::ExternalPtr;

pub use function::Function;

pub use integers::Integers;

pub use lang::Language;

pub use list::FromList;

pub use list::List;

pub use list::ListIter;

pub use logicals::Logicals;

pub use matrix::MatrixConversions;

pub use matrix::RArray;

pub use matrix::RColumn;

pub use matrix::RMatrix;

pub use matrix::RMatrix3D;

pub use nullable::Nullable;

pub use pairlist::Pairlist;

pub use pairlist::PairlistIter;

pub use primitive::Primitive;

pub use promise::Promise;

pub use raw::Raw;

pub use s4::S4;

pub use strings::Strings;

pub use symbol::Symbol;

Modules

altrep

complexes

dataframe

doubles

environment

expr

externalptr

function

integers

lang

list

logicals

macros 🔒

matrix

Wrappers for matrices with deferred arithmetic.

nullable

pairlist

A pairlist is a linked list of values with optional symbol tags.

primitive

promise

raw

rstr

s4

S4 class support.

strings

symbol

Macros

make_conversions 🔒

make_getsexp 🔒

Traits

Conversions

SymPair

Functions

make_symbol 🔒

make_vector 🔒