Module extendr_api::wrapper

Wrappers are lightweight proxies for references to R datatypes. They do not contain an Robj (see array.rs for an example of this).

Structs

Character
Wrapper for creating character objects. These are used only as the contents of a character vector.

Env
Wrapper for creating environments.

Expr
Wrapper for creating expression objects.

Func
Wrapper for creating functions (CLOSSXP).

Lang
Wrapper for creating language objects.

List
Wrapper for creating list (VECSXP) objects.

Pairlist
Wrapper for creating pair list (LISTSXP) objects.

Primitive
Wrapper for creating and reading Primitive functions.

Promise

Raw
Wrapper for creating raw (byte) objects.

Enums

Nullable
Wrapper for handling potentially NULL values.

Traits

SymPair