extendr_api::wrapper

Module altrep

Macros

make_from_iterator 🔒

make_from_iterator_impl 🔒

Structs

Altrep

Traits

AltComplexImpl

AltIntegerImpl

AltListImpl

AltLogicalImpl

AltRawImpl

AltRealImpl

AltStringImpl

AltrepImpl

Rust trait for implementing ALTREP. Implement one or more of these methods to generate an Altrep class. This is likely to be unstable for a while.

Functions

manifest 🔒