Trait extendr_api::wrapper::altrep::AltRealImpl

pub trait AltRealImpl: AltrepImpl {
    fn elt(&self, _index: usize) -> Rfloat;

    fn tot_min_max_nas(&self) -> (f64, f64, f64, usize, usize) { ... }
    fn get_region(&self, index: usize, data: &mut [Rfloat]) -> usize { ... }
    fn is_sorted(&self) -> Rbool { ... }
    fn no_na(&self) -> bool { ... }
    fn sum(&self, remove_nas: bool) -> Robj { ... }
    fn min(&self, remove_nas: bool) -> Robj { ... }
    fn max(&self, remove_nas: bool) -> Robj { ... }
}

Required methods

fn elt(&self, _index: usize) -> Rfloat

Get a single element from this vector.

Provided methods

fn tot_min_max_nas(&self) -> (f64, f64, f64, usize, usize)

fn get_region(&self, index: usize, data: &mut [Rfloat]) -> usize

Get a multiple elements from this vector.

fn is_sorted(&self) -> Rbool

Return TRUE if this vector is sorted, FALSE if not and Rbool::na() if unknown.

fn no_na(&self) -> bool

Return true if this vector does not contain NAs.

fn sum(&self, remove_nas: bool) -> Robj

Return the sum of the elements in this vector. If remove_nas is true, skip and NA values.

fn min(&self, remove_nas: bool) -> Robj

Return the minimum of the elements in this vector. If remove_nas is true, skip and NA values.

fn max(&self, remove_nas: bool) -> Robj

Return the maximum of the elements in this vector. If remove_nas is true, skip and NA values.