Trait extendr_api::wrapper::altrep::AltrepImpl

pub trait AltrepImpl: Clone + Debug {
    fn length(&self) -> usize;

    fn unserialize_ex(
        class: Robj, 
        state: Robj, 
        attributes: Robj, 
        obj_flags: i32, 
        levels: i32
    ) -> Robj { ... }

    fn unserialize(_class: Robj, _state: Robj) -> Robj { ... }

    fn serialized_state(_x: SEXP) -> Robj { ... }

    fn duplicate_ex(x: SEXP, deep: bool) -> Robj { ... }

    fn duplicate(x: SEXP, _deep: bool) -> Robj { ... }

    fn coerce(_x: SEXP, _ty: RType) -> Robj { ... }

    fn inspect(&self, _pre: i32, _deep: bool, _pvec: i32) -> bool { ... }

    fn dataptr(x: SEXP, _writeable: bool) -> *mut u8 { ... }

    fn dataptr_or_null(x: SEXP) -> *const u8 { ... }

    fn extract_subset(_x: Robj, _indx: Robj, _call: Robj) -> Robj { ... }
}

Rust trait for implementing ALTREP. Implement one or more of these methods to generate an Altrep class. This is likely to be unstable for a while.

Required methods

fn length(&self) -> usize

Get the virtual length of the vector. For example for a compact range, return end - start + 1.

Provided methods

fn unserialize_ex(
class: Robj,
state: Robj,
attributes: Robj,
obj_flags: i32,
levels: i32
) -> Robj

Constructor that is called when loading an Altrep object from a file.

fn unserialize(_class: Robj, _state: Robj) -> Robj

Simplified constructor that is called when loading an Altrep object from a file.

fn serialized_state(_x: SEXP) -> Robj

Fetch the state of this object when writing to a file.

fn duplicate_ex(x: SEXP, deep: bool) -> Robj

Duplicate this object, possibly duplicating attributes. Currently this manifests the array but preserves the original object.

fn duplicate(x: SEXP, _deep: bool) -> Robj

Duplicate this object. Called by Rf_duplicate. Currently this manifests the array but preserves the original object.

fn coerce(_x: SEXP, _ty: RType) -> Robj

Coerce this object into some other type, if possible.

fn inspect(&self, _pre: i32, _deep: bool, _pvec: i32) -> bool

Print the text for .Internal(inspect(obj))

fn dataptr(x: SEXP, _writeable: bool) -> *mut u8

Get the data pointer for this vector, possibly expanding the compact representation into a full R vector.

fn dataptr_or_null(x: SEXP) -> *const u8

Get the data pointer for this vector, returning NULL if the object is unmaterialized.

fn extract_subset(_x: Robj, _indx: Robj, _call: Robj) -> Robj

Implement subsetting (eg. x[10:19]) for this Altrep vector.