Struct extendr_api::wrapper::externalptr::ExternalPtr

pub struct ExternalPtr<T: Debug + 'static> { /* fields omitted */ }

Wrapper for creating R objects containing any Rust object.

use extendr_api::prelude::*;
test! {
    let extptr = ExternalPtr::new(1);
    assert_eq!(*extptr, 1);
    let robj : Robj = extptr.into();
    let extptr2 : ExternalPtr<i32> = robj.try_into().unwrap();
    assert_eq!(*extptr2, 1);
}

Implementations

impl<T: Any + Debug> ExternalPtr<T>

pub fn new(val: T) -> Self

Construct an external pointer object from any type T. In this case, the R object owns the data and will drop the Rust object when the last reference is removed via register_c_finalizer.

An ExternalPtr behaves like a Box except that the information is tracked by a R object.

pub fn tag(&self) -> Robj

Get the “tag” of an external pointer. This is the type name in the common case.

pub fn protected(&self) -> Robj

Get the “protected” field of an external pointer. This is NULL in the common case.

pub fn addr<'a>(&self) -> &'a T

Get the “address” field of an external pointer. Normally, we will use Deref to do this.

pub fn addr_mut(&mut self) -> &mut T

Get the “address” field of an external pointer as a mutable reference. Normally, we will use DerefMut to do this.