Struct extendr_api::wrapper::matrix::RArray

pub struct RArray<T, D> {
    robj: Robj,
    data: *mut T,
    dim: D,
}

Wrapper for creating and using matrices and arrays.

use extendr_api::prelude::*;
test! {
    let matrix = RMatrix::new_matrix(3, 2, |r, c| [
        [1., 2., 3.],
         [4., 5., 6.]][c][r]);
    let robj = r!(matrix);
    assert_eq!(robj.is_matrix(), true);
    assert_eq!(robj.nrows(), 3);
    assert_eq!(robj.ncols(), 2);

    let matrix2 : RMatrix<f64> = robj.as_matrix().ok_or("error")?;
    assert_eq!(matrix2.data().len(), 6);
    assert_eq!(matrix2.nrows(), 3);
    assert_eq!(matrix2.ncols(), 2);
}

Fields
robj: Robj

Owning Robj (probably should be a Pin).

data: *mut T

Slice of the data references the Robj.

dim: D

Dimensions of the array.

Implementations

impl<T, D> RArray<T, D>

pub fn from_parts(robj: Robj, data: *mut T, dim: D) -> Self

pub fn data(&self) -> &[T]

Get the underlying data fro this array.

pub fn dim(&self) -> &D

Get the dimensions for this array.

impl<'a, T: ToVectorValue + 'a> RArray<T, [usize; 1]>
where Robj: AsTypedSlice<'a, T>,

pub fn new_column<F: FnMut(usize) -> T>(nrows: usize, f: F) -> Self

Make a new column type.

pub fn nrows(&self) -> usize

Get the number of rows.

impl<'a, T: ToVectorValue + 'a> RArray<T, [usize; 2]>
where Robj: AsTypedSlice<'a, T>,

pub fn new_matrix<F: Clone + FnMut(usize, usize) -> T>( nrows: usize, ncols: usize, f: F ) -> Self

Create a new matrix wrapper.

Arguments

nrows - the number of rows the returned matrix will have

ncols - the number of columns the returned matrix will have

f - a function that will be called for each entry of the matrix in order to populate it with values. It must return a scalar value that can be converted to an R scalar, such as u32, f64, i.e. see ToVectorValue. It accepts two arguments:

r - the current row of the entry we are creating

c - the current column of the entry we are creating

pub fn nrows(&self) -> usize

Get the number of rows.

pub fn ncols(&self) -> usize

Get the number of columns.

impl<'a, T: ToVectorValue + 'a> RArray<T, [usize; 3]>
where Robj: AsTypedSlice<'a, T>,

pub fn new_matrix3d<F: Clone + FnMut(usize, usize, usize) -> T>( nrows: usize, ncols: usize, nmatrix: usize, f: F ) -> Self

pub fn nrows(&self) -> usize

Get the number of rows.

pub fn ncols(&self) -> usize

Get the number of columns.

pub fn nsub(&self) -> usize

Get the number of submatrices.