Type Alias extendr_api::wrapper::matrix::RMatrix

pub type RMatrix<T> = RArray<T, [usize; 2]>;

Aliased Type

struct RMatrix<T> {
    robj: Robj,
    dim: [usize; 2],
    _data: PhantomData<T>,
}

Fields
robj: Robj

Owning Robj (probably should be a Pin).

dim: [usize; 2]

Dimensions of the array.

_data: PhantomData<T>
Implementations

impl<T> RMatrix<T>
where T: ToVectorValue, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new(nrow: usize, ncol: usize) -> Self

Returns an RMatrix with dimensions according to nrow and ncol, with arbitrary entries. To initialize a matrix containing only NA values, use RMatrix::new_with_na.

impl<T> RMatrix<T>
where T: ToVectorValue + CanBeNA, Robj: for<'a> AsTypedSlice<'a, T>,

pub fn new_with_na(nrow: usize, ncol: usize) -> Self

Returns an RMatrix with dimensions according to nrow and ncol, with all entries set to NA.

Note that since Raw does not have an NA representation in R, this method is not implemented for [Rbyte].

impl<'a, T: ToVectorValue + 'a> RMatrix<T>
where Robj: AsTypedSlice<'a, T>,

pub fn new_matrix<F: Clone + FnMut(usize, usize) -> T>( nrows: usize, ncols: usize, f: F ) -> Self

Create a new matrix wrapper.

Arguments

nrows - the number of rows the returned matrix will have

ncols - the number of columns the returned matrix will have

f - a function that will be called for each entry of the matrix in order to populate it with values. It must return a scalar value that can be converted to an R scalar, such as i32, u32, or f64, i.e. see ToVectorValue. It accepts two arguments:

r - the current row of the entry we are creating

c - the current column of the entry we are creating

pub fn nrows(&self) -> usize

Get the number of rows.

pub fn ncols(&self) -> usize

Get the number of columns.